Chemical Manufacturing Process Dataset

Chemical Manufacturing Process Dataset

Usage

ChemicalManufacturingProcess

Format

A 176 x 58 array of continuous variables

Source

http://appliedpredictivemodeling.com/data

Details

This data set contains information about a chemical manufacturing process, in which the goal is to understand the relationship between the process and the resulting final product yield and can be found in the in the AppliedPredictiveModeling R package. The data has been copied from http://appliedpredictivemodeling.com/data in agreement with their license.

Examples

data(ChemicalManufacturingProcess)
head(ChemicalManufacturingProcess)
#>   Yield BiologicalMaterial01 BiologicalMaterial02 BiologicalMaterial03
#> 1 38.00                 6.25                49.58                56.97
#> 2 42.44                 8.01                60.97                67.48
#> 3 42.03                 8.01                60.97                67.48
#> 4 41.42                 8.01                60.97                67.48
#> 5 42.49                 7.47                63.33                72.25
#> 6 43.57                 6.12                58.36                65.31
#>   BiologicalMaterial04 BiologicalMaterial05 BiologicalMaterial06
#> 1                12.74                19.51                43.73
#> 2                14.65                19.36                53.14
#> 3                14.65                19.36                53.14
#> 4                14.65                19.36                53.14
#> 5                14.02                17.91                54.66
#> 6                15.17                21.79                51.23
#>   BiologicalMaterial07 BiologicalMaterial08 BiologicalMaterial09
#> 1                  100                16.66                11.44
#> 2                  100                19.04                12.55
#> 3                  100                19.04                12.55
#> 4                  100                19.04                12.55
#> 5                  100                18.22                12.80
#> 6                  100                18.30                12.13
#>   BiologicalMaterial10 BiologicalMaterial11 BiologicalMaterial12
#> 1                 3.46               138.09                18.83
#> 2                 3.46               153.67                21.05
#> 3                 3.46               153.67                21.05
#> 4                 3.46               153.67                21.05
#> 5                 3.05               147.61                21.05
#> 6                 3.78               151.88                20.76
#>   ManufacturingProcess01 ManufacturingProcess02 ManufacturingProcess03
#> 1                     NA                     NA                     NA
#> 2                    0.0                      0                     NA
#> 3                    0.0                      0                     NA
#> 4                    0.0                      0                     NA
#> 5                   10.7                      0                     NA
#> 6                   12.0                      0                     NA
#>   ManufacturingProcess04 ManufacturingProcess05 ManufacturingProcess06
#> 1                     NA                     NA                     NA
#> 2                    917                 1032.2                  210.0
#> 3                    912                 1003.6                  207.1
#> 4                    911                 1014.6                  213.3
#> 5                    918                 1027.5                  205.7
#> 6                    924                 1016.8                  208.9
#>   ManufacturingProcess07 ManufacturingProcess08 ManufacturingProcess09
#> 1                     NA                     NA                  43.00
#> 2                    177                    178                  46.57
#> 3                    178                    178                  45.07
#> 4                    177                    177                  44.92
#> 5                    178                    178                  44.96
#> 6                    178                    178                  45.32
#>   ManufacturingProcess10 ManufacturingProcess11 ManufacturingProcess12
#> 1                     NA                     NA                     NA
#> 2                     NA                     NA                      0
#> 3                     NA                     NA                      0
#> 4                     NA                     NA                      0
#> 5                     NA                     NA                      0
#> 6                     NA                     NA                      0
#>   ManufacturingProcess13 ManufacturingProcess14 ManufacturingProcess15
#> 1                   35.5                   4898                   6108
#> 2                   34.0                   4869                   6095
#> 3                   34.8                   4878                   6087
#> 4                   34.8                   4897                   6102
#> 5                   34.6                   4992                   6233
#> 6                   34.0                   4985                   6222
#>   ManufacturingProcess16 ManufacturingProcess17 ManufacturingProcess18
#> 1                   4682                   35.5                   4865
#> 2                   4617                   34.0                   4867
#> 3                   4617                   34.8                   4877
#> 4                   4635                   34.8                   4872
#> 5                   4733                   33.9                   4886
#> 6                   4786                   33.4                   4862
#>   ManufacturingProcess19 ManufacturingProcess20 ManufacturingProcess21
#> 1                   6049                   4665                    0.0
#> 2                   6097                   4621                    0.0
#> 3                   6078                   4621                    0.0
#> 4                   6073                   4611                    0.0
#> 5                   6102                   4659                   -0.7
#> 6                   6115                   4696                   -0.6
#>   ManufacturingProcess22 ManufacturingProcess23 ManufacturingProcess24
#> 1                     NA                     NA                     NA
#> 2                      3                      0                      3
#> 3                      4                      1                      4
#> 4                      5                      2                      5
#> 5                      8                      4                     18
#> 6                      9                      1                      1
#>   ManufacturingProcess25 ManufacturingProcess26 ManufacturingProcess27
#> 1                   4873                   6074                   4685
#> 2                   4869                   6107                   4630
#> 3                   4897                   6116                   4637
#> 4                   4892                   6111                   4630
#> 5                   4930                   6151                   4684
#> 6                   4871                   6128                   4687
#>   ManufacturingProcess28 ManufacturingProcess29 ManufacturingProcess30
#> 1                   10.7                   21.0                    9.9
#> 2                   11.2                   21.4                    9.9
#> 3                   11.1                   21.3                    9.4
#> 4                   11.1                   21.3                    9.4
#> 5                   11.3                   21.6                    9.0
#> 6                   11.4                   21.7                   10.1
#>   ManufacturingProcess31 ManufacturingProcess32 ManufacturingProcess33
#> 1                   69.1                    156                     66
#> 2                   68.7                    169                     66
#> 3                   69.3                    173                     66
#> 4                   69.3                    171                     68
#> 5                   69.4                    171                     70
#> 6                   68.2                    173                     70
#>   ManufacturingProcess34 ManufacturingProcess35 ManufacturingProcess36
#> 1                    2.4                    486                  0.019
#> 2                    2.6                    508                  0.019
#> 3                    2.6                    509                  0.018
#> 4                    2.5                    496                  0.018
#> 5                    2.5                    468                  0.017
#> 6                    2.5                    490                  0.018
#>   ManufacturingProcess37 ManufacturingProcess38 ManufacturingProcess39
#> 1                    0.5                      3                    7.2
#> 2                    2.0                      2                    7.2
#> 3                    0.7                      2                    7.2
#> 4                    1.2                      2                    7.2
#> 5                    0.2                      2                    7.3
#> 6                    0.4                      2                    7.2
#>   ManufacturingProcess40 ManufacturingProcess41 ManufacturingProcess42
#> 1                     NA                     NA                   11.6
#> 2                    0.1                   0.15                   11.1
#> 3                    0.0                   0.00                   12.0
#> 4                    0.0                   0.00                   10.6
#> 5                    0.0                   0.00                   11.0
#> 6                    0.0                   0.00                   11.5
#>   ManufacturingProcess43 ManufacturingProcess44 ManufacturingProcess45
#> 1                    3.0                    1.8                    2.4
#> 2                    0.9                    1.9                    2.2
#> 3                    1.0                    1.8                    2.3
#> 4                    1.1                    1.8                    2.1
#> 5                    1.1                    1.7                    2.1
#> 6                    2.2                    1.8                    2.0