Australian Vacation Activities

The Australian Vacation Activities dataset includes responses from 1,003 adult Australians who were surveyed on their vacation activities through a permission-based internet panel in 2007.The responses are coded in binary with 0 indicating that the tourist didn't partake in the activity and 1 indicating they did.

Usage

ausActiv

Format

A 1003 x 44 array of binary responses

Examples

head(ausActiv)
#>      Bushwalk Beach Farm Whale Gardens Camping Swimming Skiing Tennis Riding
#> [1,]        0     0    1     0       0       1        0      0      0      0
#> [2,]        0     0    0     0       0       0        0      0      1      0
#> [3,]        0     1    0     1       1       0        1      1      1      0
#> [4,]        0     1    0     0       0       0        0      0      0      0
#> [5,]        0     0    0     0       0       0        0      0      0      0
#> [6,]        0     1    0     0       0       0        1      0      0      0
#>      Cycling Hiking Exercising Golf Fishing ScubaDiving Surfing FourWheel
#> [1,]       0      0          0    0       0           0       0         0
#> [2,]       0      0          1    0       0           0       0         0
#> [3,]       1      0          1    0       1           1       1         0
#> [4,]       0      0          0    0       1           0       0         0
#> [5,]       0      0          0    0       0           0       0         0
#> [6,]       0      0          0    0       0           0       0         0
#>      Adventure WaterSport Theatre Monuments Cultural Festivals Museum ThemePark
#> [1,]         0          0       0         0        0         0      0         0
#> [2,]         0          0       0         0        0         0      0         0
#> [3,]         0          1       1         1        1         0      0         1
#> [4,]         0          0       0         0        0         0      0         0
#> [5,]         0          0       0         0        0         0      0         0
#> [6,]         0          0       0         0        0         0      0         0
#>      CharterBoat Spa ScenicWalks Markets GuidedTours Industrial Wildlife
#> [1,]           0   0           1       0           0          0        0
#> [2,]           0   0           0       0           0          0        0
#> [3,]           0   0           1       0           0          1        1
#> [4,]           0   0           0       0           0          0        0
#> [5,]           0   0           1       1           0          0        0
#> [6,]           1   0           1       1           0          0        0
#>      ChildrenAtt Sightseeing Friends Pubs BBQ Shopping Eating EatingHigh Movies
#> [1,]           0           1       0    0   0        0      0          0      0
#> [2,]           0           0       1    0   0        1      1          1      1
#> [3,]           0           0       0    0   1        1      1          0      0
#> [4,]           0           0       0    0   0        0      0          0      0
#> [5,]           0           1       1    0   0        1      1          0      0
#> [6,]           0           1       0    1   0        1      1          1      0
#>      Casino Relaxing SportEvent
#> [1,]      0        0          0
#> [2,]      0        1          0
#> [3,]      0        0          0
#> [4,]      0        1          0
#> [5,]      0        0          0
#> [6,]      0        1          0