Get Guided Tour-Holes-Better History

Returns a guided tour with the holes index and the 'search_better' argument to the python' backend. This guided tour is generated with the 'tourr' functions 'save_history' and 'guided_tour'.

Usage

get_guided_holes_better_history(data, dimension)

Arguments

data

the dataset to calculate the projections with.

dimension

1 for a 1d tour or 2 for a 2d tour

Value

history object containing the projections of the requested tour

Examples

data("flea", package = "tourr")
flea <- flea[-7]
get_guided_holes_better_history(flea, 2)
#> Converting input data to the required matrix format.
#> Target: 0.893, try: 1
#> Target: 1.044, try: 1
#> Target: 1.057, try: 11
#> Target: 1.172, try: 3
#> Target: 1.233, try: 19
#> No better bases found after 25 tries.  Giving up.
#> Final projection: 
#>  0.438  -0.107  
#> -0.138  -0.360  
#> -0.494   0.053  
#> -0.310  -0.624  
#>  0.586  -0.529  
#> -0.324  -0.433  
#> , , 1
#> 
#>            [,1]       [,2]
#> [1,]  0.5062877  0.3267385
#> [2,] -0.8031649  0.4228485
#> [3,] -0.1090356 -0.5299303
#> [4,] -0.1076576  0.1290855
#> [5,] -0.1246333 -0.6053402
#> [6,] -0.2441034 -0.2247557
#> 
#> , , 2
#> 
#>             [,1]       [,2]
#> [1,]  0.35124505  0.3970632
#> [2,] -0.36754817  0.7290904
#> [3,] -0.45756704 -0.1898052
#> [4,]  0.11390589 -0.2610223
#> [5,]  0.05798261 -0.4342012
#> [6,] -0.71821384 -0.1344570
#> 
#> , , 3
#> 
#>            [,1]        [,2]
#> [1,] -0.3530215  0.58108411
#> [2,] -0.1920326 -0.31578667
#> [3,] -0.2312760 -0.50557530
#> [4,]  0.2461555 -0.43848705
#> [5,]  0.7114820 -0.03101622
#> [6,] -0.4671312 -0.33731404
#> 
#> , , 4
#> 
#>            [,1]        [,2]
#> [1,]  0.2997709  0.53737844
#> [2,] -0.3106101 -0.17829413
#> [3,] -0.2625418  0.01524148
#> [4,] -0.5160806 -0.34100634
#> [5,]  0.1530247 -0.60941250
#> [6,] -0.6745182  0.43764644
#> 
#> , , 5
#> 
#>             [,1]        [,2]
#> [1,]  0.31255102  0.39881926
#> [2,]  0.22045685 -0.57241740
#> [3,]  0.01051489  0.01629131
#> [4,] -0.30804031 -0.53235298
#> [5,]  0.66763140 -0.42654184
#> [6,] -0.55944576 -0.21835401
#> 
#> , , 6
#> 
#>            [,1]         [,2]
#> [1,]  0.4468936 -0.063439408
#> [2,] -0.1019289 -0.371729974
#> [3,] -0.4970908  0.003744754
#> [4,] -0.2462942 -0.651179551
#> [5,]  0.6355691 -0.468200416
#> [6,] -0.2796218 -0.463175785
#> 
#> , , 7
#> 
#>            [,1]         [,2]
#> [1,]  0.4468936 -0.063439408
#> [2,] -0.1019289 -0.371729974
#> [3,] -0.4970908  0.003744754
#> [4,] -0.2462942 -0.651179551
#> [5,]  0.6355691 -0.468200416
#> [6,] -0.2796218 -0.463175785
#> 
#> attr(,"class")
#> [1] "history_array"