Get Guided Tour-Holes History

Returns a guided tour with the holes index to the python' backend. This guided tour is generated with the 'tourr' functions 'save_history' and 'guided_tour'.

Usage

get_guided_holes_history(data, dimension)

Arguments

data

the dataset to calculate the projections with.

dimension

1 for a 1d tour or 2 for a 2d tour

Value

history object containing the projections of the requested tour

Examples

data("flea", package = "tourr")
flea <- flea[-7]
get_guided_holes_history(flea, 2)
#> Converting input data to the required matrix format.
#> Target: 1.003, 49.2% better 
#> Target: 1.079, 7.6% better 
#> Target: 1.090, 0.9% better 
#> Target: 1.125, 3.3% better 
#> Target: 1.161, 3.2% better 
#> Target: 1.180, 1.6% better 
#> Target: 1.214, 2.9% better 
#> Target: 1.246, 2.6% better 
#> Target: 1.259, 1.0% better 
#> Target: 1.267, 0.7% better 
#> Target: 1.271, 0.3% better 
#> Target: 1.274, 0.2% better 
#> Target: 1.275, 0.1% better 
#> Target: 1.276, 0.1% better 
#> Target: 1.278, 0.1% better 
#> No better bases found after 25 tries.  Giving up.
#> Final projection: 
#>  0.655  -0.258  
#>  0.536   0.391  
#>  0.190   0.363  
#>  0.063   0.621  
#>  0.474  -0.177  
#> -0.135   0.481  
#> , , 1
#> 
#>            [,1]        [,2]
#> [1,]  0.2412202 -0.56692762
#> [2,]  0.2769358  0.24237708
#> [3,] -0.1671454  0.62350237
#> [4,]  0.5100804  0.31826170
#> [5,] -0.1700304  0.34743510
#> [6,] -0.7403306 -0.09533916
#> 
#> , , 2
#> 
#>             [,1]       [,2]
#> [1,]  0.30902277 -0.6064381
#> [2,] -0.01767911  0.3050899
#> [3,]  0.01574986  0.2186512
#> [4,]  0.53821879  0.6611782
#> [5,] -0.02748435  0.1025455
#> [6,] -0.78326845  0.2089794
#> 
#> , , 3
#> 
#>             [,1]       [,2]
#> [1,]  0.22658679 -0.5741958
#> [2,]  0.08512613  0.2765525
#> [3,] -0.14327966  0.1518313
#> [4,]  0.49476421  0.6767873
#> [5,]  0.15650842 -0.2665210
#> [6,] -0.80721521  0.2041831
#> 
#> , , 4
#> 
#>            [,1]       [,2]
#> [1,]  0.1724696 -0.5295564
#> [2,]  0.0735436  0.2385948
#> [3,] -0.2083910  0.1537659
#> [4,]  0.4853975  0.7369179
#> [5,]  0.2174087 -0.1365156
#> [6,] -0.7990879  0.2780543
#> 
#> , , 5
#> 
#>            [,1]       [,2]
#> [1,]  0.3260132 -0.4216591
#> [2,]  0.1049052  0.4082176
#> [3,] -0.2631617  0.1168955
#> [4,]  0.5615840  0.6801095
#> [5,]  0.3259026 -0.3606032
#> [6,] -0.6259929  0.2220670
#> 
#> , , 6
#> 
#>             [,1]       [,2]
#> [1,]  0.48034399 -0.2890758
#> [2,]  0.03655215  0.5112531
#> [3,] -0.43136757  0.2006853
#> [4,]  0.60227754  0.6584276
#> [5,]  0.35000723 -0.2170986
#> [6,] -0.31082522  0.3662269
#> 
#> , , 7
#> 
#>             [,1]        [,2]
#> [1,]  0.41446480 -0.40956348
#> [2,]  0.07404912  0.42625938
#> [3,] -0.29538299  0.09314089
#> [4,]  0.60127997  0.67836049
#> [5,]  0.44406204 -0.17919574
#> [6,] -0.42042341  0.38678340
#> 
#> , , 8
#> 
#>             [,1]       [,2]
#> [1,]  0.55726176 -0.4137289
#> [2,]  0.38650028  0.4052031
#> [3,] -0.09799995  0.3987372
#> [4,]  0.42921195  0.6491274
#> [5,]  0.33581473 -0.1020349
#> [6,] -0.48319607  0.2717903
#> 
#> , , 9
#> 
#>             [,1]       [,2]
#> [1,]  0.56002314 -0.3671667
#> [2,]  0.58700784  0.4621414
#> [3,] -0.04353609  0.1863137
#> [4,]  0.22827835  0.6696795
#> [5,]  0.41298020 -0.2248906
#> [6,] -0.34239865  0.3432996
#> 
#> , , 10
#> 
#>            [,1]       [,2]
#> [1,]  0.6527261 -0.2987868
#> [2,]  0.5957880  0.4201676
#> [3,] -0.0297502  0.3027087
#> [4,]  0.1994317  0.6456473
#> [5,]  0.3142699 -0.1535334
#> [6,] -0.2820667  0.4495776
#> 
#> , , 11
#> 
#>            [,1]       [,2]
#> [1,]  0.6651624 -0.3005131
#> [2,]  0.5460540  0.4356398
#> [3,]  0.1657625  0.3920781
#> [4,]  0.1574118  0.6019429
#> [5,]  0.3278059 -0.2277682
#> [6,] -0.3157081  0.3898346
#> 
#> , , 12
#> 
#>            [,1]       [,2]
#> [1,]  0.6716792 -0.2599205
#> [2,]  0.5325854  0.4587768
#> [3,]  0.1782011  0.3084536
#> [4,]  0.1473420  0.6043457
#> [5,]  0.3768077 -0.2576344
#> [6,] -0.2641032  0.4418283
#> 
#> , , 13
#> 
#>            [,1]       [,2]
#> [1,]  0.6563003 -0.2503390
#> [2,]  0.5371260  0.4031523
#> [3,]  0.1396529  0.3457557
#> [4,]  0.1589155  0.6391943
#> [5,]  0.4222766 -0.2274899
#> [6,] -0.2401897  0.4415095
#> 
#> , , 14
#> 
#>            [,1]       [,2]
#> [1,]  0.6449292 -0.2554972
#> [2,]  0.5615712  0.3910975
#> [3,]  0.1051530  0.3746676
#> [4,]  0.1305985  0.6191378
#> [5,]  0.4534029 -0.1921212
#> [6,] -0.1871279  0.4702615
#> 
#> , , 15
#> 
#>            [,1]       [,2]
#> [1,]  0.6559170 -0.2584684
#> [2,]  0.5350638  0.3754991
#> [3,]  0.1700302  0.3764765
#> [4,]  0.1076885  0.6085564
#> [5,]  0.4590137 -0.1521865
#> [6,] -0.1796636  0.5069106
#> 
#> , , 16
#> 
#>             [,1]       [,2]
#> [1,]  0.65517254 -0.2577021
#> [2,]  0.53607625  0.3912913
#> [3,]  0.19011999  0.3631648
#> [4,]  0.06299139  0.6211876
#> [5,]  0.47434067 -0.1773296
#> [6,] -0.13512435  0.4809077
#> 
#> , , 17
#> 
#>             [,1]       [,2]
#> [1,]  0.65517254 -0.2577021
#> [2,]  0.53607625  0.3912913
#> [3,]  0.19011999  0.3631648
#> [4,]  0.06299139  0.6211876
#> [5,]  0.47434067 -0.1773296
#> [6,] -0.13512435  0.4809077
#> 
#> attr(,"class")
#> [1] "history_array"