Get Local Tour History

Get Local Tour History

Usage

get_local_history(data, starting_projection)

Arguments

data

the dataset to calculate the projections with. In practice only the two first rows of the dataset are provided as the actual data is not needed.

starting_projection

the initial projection one wants to initiate the local tour from

Value

history object containing the projections of the requested tour

Details

Function for python backend to receive a local tour based on current projection

Examples

library(tourr)
data(flea)
flea <- flea[-7]
prj <- basis_random(ncol(flea), 2)
get_local_history(flea, prj)
#> Converting input data to the required matrix format.
#> , , 1
#> 
#>              [,1]        [,2]
#> [1,] -0.454694007  0.75953637
#> [2,] -0.002655978 -0.28657496
#> [3,] -0.239897858 -0.40343779
#> [4,]  0.480197766  0.09630225
#> [5,]  0.710451430  0.29150136
#> [6,]  0.019083879 -0.28977590
#> 
#> , , 2
#> 
#>             [,1]        [,2]
#> [1,] -0.36624052  0.85118438
#> [2,]  0.09871645  0.04484003
#> [3,] -0.01679317 -0.21026956
#> [4,]  0.58944778  0.05652186
#> [5,]  0.71241146  0.39080724
#> [6,] -0.02936257  0.27080658
#> 
#> , , 3
#> 
#>            [,1]       [,2]
#> [1,] -0.5190146  0.6848185
#> [2,] -0.1819157 -0.3271894
#> [3,] -0.3214417 -0.6098694
#> [4,]  0.3999555  0.1373338
#> [5,]  0.6432744  0.1032178
#> [6,]  0.1429661 -0.1500521
#> 
#> , , 4
#> 
#>             [,1]         [,2]
#> [1,] -0.49726273  0.730764460
#> [2,] -0.28901654 -0.480296651
#> [3,] -0.41197910  0.004604299
#> [4,]  0.61330890  0.226316992
#> [5,]  0.34027153  0.328149937
#> [6,]  0.08683266 -0.276361123
#> 
#> , , 5
#> 
#>             [,1]       [,2]
#> [1,] -0.48456758  0.7204013
#> [2,] -0.05773918 -0.3133021
#> [3,] -0.21881268 -0.4676318
#> [4,]  0.53365826 -0.0327377
#> [5,]  0.64476875  0.3435331
#> [6,] -0.11603265 -0.2123618
#> 
#> , , 6
#> 
#>             [,1]       [,2]
#> [1,] -0.46283935  0.7873403
#> [2,]  0.05730548 -0.2482572
#> [3,] -0.29371180 -0.4065317
#> [4,]  0.49626518  0.1868523
#> [5,]  0.67062978  0.2430807
#> [6,] -0.01434472 -0.2432974
#> 
#> , , 7
#> 
#>             [,1]       [,2]
#> [1,] -0.46039991  0.7577948
#> [2,]  0.03658077 -0.2728549
#> [3,] -0.24184373 -0.4775725
#> [4,]  0.44122497  0.1229632
#> [5,]  0.72960837  0.2490216
#> [6,] -0.03460552 -0.2146859
#> 
#> , , 8
#> 
#>              [,1]       [,2]
#> [1,] -0.303408339  0.8254026
#> [2,]  0.048154736 -0.1292563
#> [3,] -0.002337504 -0.3812604
#> [4,]  0.641385577  0.1790362
#> [5,]  0.688862705  0.2544511
#> [6,]  0.140398547 -0.2446315
#> 
#> , , 9
#> 
#>              [,1]       [,2]
#> [1,] -0.487555898  0.7077792
#> [2,]  0.005689836 -0.3275232
#> [3,] -0.252317533 -0.4389803
#> [4,]  0.365281508  0.1063738
#> [5,]  0.750092363  0.2388572
#> [6,] -0.050235461 -0.3615319
#> 
#> , , 10
#> 
#>            [,1]       [,2]
#> [1,] -0.4890934  0.7371789
#> [2,] -0.1359154 -0.1552228
#> [3,] -0.1342308 -0.5788169
#> [4,]  0.4713979  0.2013662
#> [5,]  0.6461160  0.1990102
#> [6,] -0.2908863 -0.1314940
#> 
#> , , 11
#> 
#>            [,1]       [,2]
#> [1,] -0.3777836  0.5231728
#> [2,]  0.2036565 -0.4766027
#> [3,] -0.1253391 -0.4899734
#> [4,]  0.5449412  0.0422768
#> [5,]  0.6815176  0.1707011
#> [6,] -0.1966383 -0.4776401
#> 
#> , , 12
#> 
#>            [,1]       [,2]
#> [1,] -0.3750040  0.7855379
#> [2,]  0.1073443 -0.1602953
#> [3,] -0.2330141 -0.4695991
#> [4,]  0.3914598  0.1304598
#> [5,]  0.7796577  0.2494389
#> [6,]  0.1801297 -0.2397346
#> 
#> , , 13
#> 
#>               [,1]       [,2]
#> [1,] -0.4247872596  0.7242115
#> [2,] -0.1220783014 -0.4094791
#> [3,] -0.0003074088 -0.4601151
#> [4,]  0.6416937276  0.1650694
#> [5,]  0.5999802783  0.2019609
#> [6,] -0.1813984662 -0.1676381
#> 
#> , , 14
#> 
#>             [,1]       [,2]
#> [1,] -0.53667775  0.7622819
#> [2,]  0.07906247 -0.1196209
#> [3,] -0.29511955 -0.2031735
#> [4,]  0.38307647  0.3255843
#> [5,]  0.64351314  0.1873512
#> [6,] -0.24036184 -0.4714149
#> 
#> , , 15
#> 
#>             [,1]        [,2]
#> [1,] -0.49546167  0.83425216
#> [2,] -0.07361689  0.06210735
#> [3,] -0.08485113 -0.22458129
#> [4,]  0.46507080  0.10462795
#> [5,]  0.65275204  0.40276586
#> [6,] -0.31547187 -0.27669822
#> 
#> , , 16
#> 
#>              [,1]        [,2]
#> [1,] -0.335122105  0.82877087
#> [2,]  0.263701748 -0.24267175
#> [3,] -0.192106278 -0.32296214
#> [4,]  0.614412743  0.08215599
#> [5,]  0.635370948  0.35942912
#> [6,] -0.007104952 -0.11834623
#> 
#> , , 17
#> 
#>             [,1]       [,2]
#> [1,] -0.47316513  0.7373732
#> [2,] -0.07219374 -0.2264887
#> [3,] -0.43970183 -0.3681418
#> [4,]  0.49288491  0.1625203
#> [5,]  0.56205755  0.2626975
#> [6,]  0.13682434 -0.4171720
#> 
#> , , 18
#> 
#>             [,1]         [,2]
#> [1,] -0.31004787  0.721215852
#> [2,]  0.14563155 -0.411120139
#> [3,] -0.12458835 -0.366192084
#> [4,]  0.64124008 -0.006634584
#> [5,]  0.67380479  0.372098314
#> [6,]  0.04402005 -0.195525219
#> 
#> , , 19
#> 
#>               [,1]        [,2]
#> [1,] -0.5059124132  0.70104800
#> [2,]  0.0007011094 -0.29440535
#> [3,] -0.3367552827 -0.41651278
#> [4,]  0.4628134573  0.04128876
#> [5,]  0.6443990051  0.28137274
#> [6,] -0.0346647388 -0.40926632
#> 
#> , , 20
#> 
#>             [,1]       [,2]
#> [1,] -0.37360712  0.8166263
#> [2,]  0.05927789 -0.1113255
#> [3,] -0.09046509 -0.2995533
#> [4,]  0.17651757  0.1642090
#> [5,]  0.90117248  0.3105157
#> [6,]  0.07382164 -0.3280415
#> 
#> , , 21
#> 
#>             [,1]       [,2]
#> [1,] -0.49450200  0.7532189
#> [2,] -0.24693567 -0.2709981
#> [3,] -0.25371938 -0.2748825
#> [4,]  0.31744778  0.2781150
#> [5,]  0.72453678  0.2389321
#> [6,]  0.06625992 -0.3862958
#> 
#> , , 22
#> 
#>             [,1]       [,2]
#> [1,] -0.45743221  0.7285817
#> [2,]  0.05747483 -0.3677737
#> [3,] -0.28288123 -0.3390492
#> [4,]  0.50637313  0.1006921
#> [5,]  0.67103198  0.2951749
#> [6,] -0.02707330 -0.3488406
#> 
#> , , 23
#> 
#>              [,1]        [,2]
#> [1,] -0.454836459  0.75928556
#> [2,] -0.001890646 -0.28916360
#> [3,] -0.242283960 -0.40308040
#> [4,]  0.477143340  0.09626462
#> [5,]  0.711545238  0.29144088
#> [6,]  0.021361426 -0.28842917
#> 
#> , , 24
#> 
#>            [,1]        [,2]
#> [1,] -0.2947454  0.63739540
#> [2,]  0.1326216 -0.44778912
#> [3,] -0.4178968 -0.33190698
#> [4,]  0.3842631  0.06315215
#> [5,]  0.6801614  0.32685583
#> [6,]  0.3325977 -0.41500222
#> 
#> , , 25
#> 
#>             [,1]        [,2]
#> [1,] -0.42849052  0.72966395
#> [2,] -0.05616801 -0.36076086
#> [3,] -0.20732802 -0.34583861
#> [4,]  0.49583636  0.04990046
#> [5,]  0.71613139  0.32355754
#> [6,]  0.10750935 -0.33265333
#> 
#> , , 26
#> 
#>              [,1]          [,2]
#> [1,] -0.484179524  0.7137004736
#> [2,] -0.005683411 -0.2651014188
#> [3,] -0.311177543 -0.4655263307
#> [4,]  0.499308190 -0.0002542948
#> [5,]  0.647583401  0.3048137581
#> [6,] -0.005787352 -0.3327560900
#> 
#> , , 27
#> 
#>             [,1]       [,2]
#> [1,] -0.43191434  0.7555117
#> [2,]  0.02272611 -0.2293075
#> [3,] -0.13676761 -0.4466630
#> [4,]  0.52371679  0.1137337
#> [5,]  0.71768691  0.2639162
#> [6,] -0.06981674 -0.3074495
#> 
#> , , 28
#> 
#>             [,1]        [,2]
#> [1,] -0.44930867  0.71508875
#> [2,]  0.03678901 -0.38010156
#> [3,] -0.25709658 -0.43139941
#> [4,]  0.44007085  0.04455273
#> [5,]  0.73239513  0.28863488
#> [6,]  0.02458968 -0.26975988
#> 
#> , , 29
#> 
#>             [,1]       [,2]
#> [1,] -0.42885063  0.7422549
#> [2,] -0.04558174 -0.4226898
#> [3,] -0.11974724 -0.3628767
#> [4,]  0.33693021  0.1225008
#> [5,]  0.82730539  0.2703201
#> [6,]  0.04139894 -0.2250159
#> 
#> , , 30
#> 
#>             [,1]        [,2]
#> [1,] -0.46185630  0.82323255
#> [2,] -0.04744381  0.03166652
#> [3,] -0.35234591 -0.19431832
#> [4,]  0.59876670  0.35055784
#> [5,]  0.54447139  0.23345384
#> [6,]  0.07293529 -0.32578260
#> 
#> , , 31
#> 
#>             [,1]        [,2]
#> [1,] -0.30543638  0.80221010
#> [2,]  0.27474109 -0.33150513
#> [3,] -0.06665886 -0.30547683
#> [4,]  0.30991000  0.03066841
#> [5,]  0.85101903  0.34203411
#> [6,]  0.08065321  0.18793436
#> 
#> , , 32
#> 
#>             [,1]        [,2]
#> [1,] -0.41045328  0.67233756
#> [2,] -0.16649473 -0.36463823
#> [3,] -0.05364514 -0.33434836
#> [4,]  0.53283803 -0.02507846
#> [5,]  0.71525987  0.33911549
#> [6,]  0.07359866 -0.43310979
#> 
#> , , 33
#> 
#>             [,1]        [,2]
#> [1,] -0.42031484  0.76856215
#> [2,]  0.10881972 -0.31561119
#> [3,] -0.22391949 -0.40600541
#> [4,]  0.55378732  0.03353368
#> [5,]  0.66885414  0.35293673
#> [6,] -0.08548402 -0.13846502
#> 
#> , , 34
#> 
#>            [,1]        [,2]
#> [1,] -0.4627079  0.75066203
#> [2,]  0.1581206 -0.37136101
#> [3,] -0.4878042 -0.25104146
#> [4,]  0.2938663  0.02042117
#> [5,]  0.6351374  0.47061491
#> [6,] -0.1821797  0.11696275
#> 
#> , , 35
#> 
#>             [,1]       [,2]
#> [1,] -0.40142062  0.7175819
#> [2,] -0.02676857 -0.1376555
#> [3,]  0.21066017 -0.1943808
#> [4,]  0.43387225 -0.1793960
#> [5,]  0.75959795  0.4281465
#> [6,] -0.16891729 -0.4613578
#> 
#> , , 36
#> 
#>            [,1]         [,2]
#> [1,] -0.2245333  0.696191621
#> [2,]  0.4416003 -0.199780737
#> [3,] -0.1771569 -0.500557172
#> [4,]  0.4748969  0.009782931
#> [5,]  0.6983257  0.272142595
#> [6,]  0.1000179 -0.388188234
#> 
#> , , 37
#> 
#>             [,1]       [,2]
#> [1,] -0.41876673  0.7801716
#> [2,] -0.11577576 -0.1647756
#> [3,] -0.11261005 -0.4058608
#> [4,]  0.39132110  0.1992175
#> [5,]  0.80088387  0.2540837
#> [6,]  0.06326293 -0.3085648
#> 
#> , , 38
#> 
#>              [,1]        [,2]
#> [1,] -0.444179904  0.74576313
#> [2,]  0.014576027 -0.39132516
#> [3,] -0.220860877 -0.28795603
#> [4,]  0.496710401  0.01744964
#> [5,]  0.711988698  0.37465363
#> [6,]  0.007943346 -0.25906268
#> 
#> , , 39
#> 
#>             [,1]       [,2]
#> [1,] -0.43067388  0.7924625
#> [2,] -0.04230972 -0.2601363
#> [3,] -0.13419435 -0.2914238
#> [4,]  0.49563378  0.1053490
#> [5,]  0.74032694  0.3358823
#> [6,]  0.03138395 -0.3090132
#> 
#> , , 40
#> 
#>             [,1]       [,2]
#> [1,] -0.23348604  0.7592694
#> [2,]  0.04492469 -0.1040060
#> [3,] -0.01423777 -0.4053129
#> [4,]  0.61203253  0.1670120
#> [5,]  0.70663435  0.2522756
#> [6,]  0.26333898 -0.3960784
#> 
#> , , 41
#> 
#>             [,1]        [,2]
#> [1,] -0.47697044  0.81610081
#> [2,] -0.08082988 -0.16389466
#> [3,] -0.39325387 -0.13727870
#> [4,]  0.54088245  0.05800846
#> [5,]  0.56354943  0.50527718
#> [6,] -0.03428328 -0.17205392
#> 
#> , , 42
#> 
#>            [,1]        [,2]
#> [1,] -0.4486542  0.57064027
#> [2,] -0.2498208 -0.45246161
#> [3,] -0.3108592 -0.37054451
#> [4,]  0.6562001  0.05788612
#> [5,]  0.4216456  0.20115911
#> [6,]  0.1768672 -0.53714910
#> 
#> , , 43
#> 
#>             [,1]        [,2]
#> [1,] -0.41916339  0.64687342
#> [2,] -0.01838471 -0.58681855
#> [3,] -0.20864795 -0.08743784
#> [4,]  0.70884990  0.04599456
#> [5,]  0.52467003  0.36955720
#> [6,] -0.05180027 -0.30143883
#> 
#> , , 44
#> 
#>             [,1]        [,2]
#> [1,] -0.44478448  0.76004757
#> [2,] -0.01142685 -0.25919548
#> [3,] -0.23941201 -0.41260788
#> [4,]  0.47586219  0.07376313
#> [5,]  0.71754479  0.30436981
#> [6,]  0.05833298 -0.29464920
#> 
#> , , 45
#> 
#>             [,1]       [,2]
#> [1,] -0.45457974  0.7126553
#> [2,] -0.03566254 -0.3484773
#> [3,]  0.01280376 -0.4329531
#> [4,]  0.69490353  0.2827194
#> [5,]  0.46881902  0.0555446
#> [6,] -0.29872951 -0.3165789
#> 
#> , , 46
#> 
#>            [,1]        [,2]
#> [1,] -0.4141159  0.84889385
#> [2,]  0.2654509 -0.03311026
#> [3,] -0.1896764 -0.28724066
#> [4,]  0.3550998  0.37945092
#> [5,]  0.7361146  0.19577396
#> [6,] -0.2326073 -0.11604012
#> 
#> , , 47
#> 
#>             [,1]       [,2]
#> [1,] -0.34355988  0.5644875
#> [2,]  0.02871997 -0.2132273
#> [3,] -0.31479076 -0.4699950
#> [4,] -0.06031276  0.2280722
#> [5,]  0.87290455  0.1607640
#> [6,]  0.12825198 -0.5806296
#> 
#> , , 48
#> 
#>             [,1]        [,2]
#> [1,] -0.51581241  0.65172302
#> [2,] -0.16215896 -0.53320752
#> [3,] -0.23548379 -0.20704964
#> [4,]  0.41383454  0.08581965
#> [5,]  0.69191707  0.26687810
#> [6,]  0.04670277 -0.41168963
#> 
#> , , 49
#> 
#>             [,1]        [,2]
#> [1,] -0.42816867  0.77739614
#> [2,]  0.09828573  0.01872918
#> [3,] -0.11002675 -0.46951053
#> [4,]  0.48689700  0.17383706
#> [5,]  0.73712048  0.21266701
#> [6,] -0.12037581 -0.31530580
#> 
#> , , 50
#> 
#>            [,1]       [,2]
#> [1,] -0.4054935  0.7920192
#> [2,] -0.1622788  0.1253791
#> [3,] -0.3727220 -0.3581413
#> [4,]  0.6671777  0.2451726
#> [5,]  0.4286194  0.2561759
#> [6,]  0.2036623 -0.3209126
#> 
#> , , 51
#> 
#>             [,1]       [,2]
#> [1,] -0.33784374  0.7021088
#> [2,]  0.03727815 -0.3476735
#> [3,] -0.10656473 -0.3449987
#> [4,]  0.53147357  0.1689240
#> [5,]  0.75639253  0.2399856
#> [6,]  0.13609591 -0.4254572
#> 
#> , , 52
#> 
#>             [,1]        [,2]
#> [1,] -0.46244509  0.71513896
#> [2,] -0.04286010 -0.29716071
#> [3,] -0.20261700 -0.45017172
#> [4,]  0.41160288  0.03020295
#> [5,]  0.75732496  0.29062146
#> [6,]  0.01720088 -0.33502860
#> 
#> , , 53
#> 
#>             [,1]        [,2]
#> [1,] -0.40097963  0.76823999
#> [2,]  0.14876017 -0.34362457
#> [3,] -0.18908708 -0.29017621
#> [4,]  0.45008483  0.03450635
#> [5,]  0.76059303  0.37426954
#> [6,] -0.01592834 -0.25740796
#> 
#> , , 54
#> 
#>              [,1]        [,2]
#> [1,] -0.433793331  0.69127752
#> [2,]  0.021631617 -0.36967307
#> [3,] -0.193632610 -0.42832893
#> [4,]  0.512409165  0.05356888
#> [5,]  0.715018429  0.27279156
#> [6,] -0.006879413 -0.35316664
#> 
#> , , 55
#> 
#>             [,1]        [,2]
#> [1,] -0.52796078  0.74408974
#> [2,] -0.03070435 -0.25754908
#> [3,] -0.38827171 -0.22001916
#> [4,]  0.33539227 -0.07728636
#> [5,]  0.67529974  0.46618363
#> [6,] -0.03228045 -0.32907467
#> 
#> , , 56
#> 
#>            [,1]         [,2]
#> [1,] -0.4047266  0.746134013
#> [2,]  0.1042560 -0.343104327
#> [3,] -0.1000582 -0.318173660
#> [4,]  0.5746217 -0.008515998
#> [5,]  0.6732685  0.390739794
#> [6,] -0.1784231 -0.267542276
#> 
#> , , 57
#> 
#>             [,1]       [,2]
#> [1,] -0.45783385  0.7032573
#> [2,] -0.19563154 -0.2829107
#> [3,] -0.15940485 -0.2785023
#> [4,]  0.43288137  0.1387930
#> [5,]  0.72971475  0.2789543
#> [6,]  0.08268409 -0.5007476
#> 
#> , , 58
#> 
#>              [,1]       [,2]
#> [1,] -0.449639072  0.7790279
#> [2,]  0.033351465 -0.2414029
#> [3,] -0.239489158 -0.4277238
#> [4,]  0.477812471  0.1078337
#> [5,]  0.714878409  0.2862891
#> [6,]  0.001196267 -0.2414601
#> 
#> , , 59
#> 
#>            [,1]        [,2]
#> [1,] -0.4264938  0.78218070
#> [2,]  0.1856131 -0.26115607
#> [3,] -0.4260505  0.05598887
#> [4,]  0.5541212  0.25884812
#> [5,]  0.4109454  0.44919419
#> [6,] -0.3552540 -0.21926775
#> 
#> , , 60
#> 
#>            [,1]        [,2]
#> [1,] -0.4884138  0.51695743
#> [2,]  0.2240314 -0.31615909
#> [3,] -0.3316003 -0.78550557
#> [4,]  0.2850069  0.06922757
#> [5,]  0.6646161  0.02167734
#> [6,] -0.2799279 -0.10255278
#> 
#> , , 61
#> 
#>               [,1]       [,2]
#> [1,] -0.4497738117  0.7574708
#> [2,] -0.0003679765 -0.2896643
#> [3,] -0.2269193531 -0.4115111
#> [4,]  0.4884952993  0.1034701
#> [5,]  0.7122672763  0.2825863
#> [6,]  0.0160829030 -0.2871065
#> 
#> , , 62
#> 
#>              [,1]        [,2]
#> [1,] -0.512625418  0.65791838
#> [2,]  0.008981921 -0.35344295
#> [3,] -0.298491190 -0.36808864
#> [4,]  0.421969540 -0.09230731
#> [5,]  0.681974032  0.35243002
#> [6,] -0.069933131 -0.41713868
#> 
#> , , 63
#> 
#>             [,1]        [,2]
#> [1,] -0.47473605  0.69812930
#> [2,] -0.17972335 -0.35129528
#> [3,] -0.19605446 -0.39475642
#> [4,]  0.50867829  0.03944774
#> [5,]  0.66468993  0.29031320
#> [6,]  0.05763281 -0.38410492
#> 
#> , , 64
#> 
#>             [,1]        [,2]
#> [1,] -0.43910008  0.77651382
#> [2,]  0.10808917 -0.28030523
#> [3,] -0.28390127 -0.33695316
#> [4,]  0.49972165  0.06366211
#> [5,]  0.68179031  0.34990966
#> [6,] -0.01865901 -0.28005036
#> 
#> , , 65
#> 
#>             [,1]        [,2]
#> [1,] -0.37560156  0.82208913
#> [2,] -0.01676289 -0.19981526
#> [3,] -0.09970850 -0.43243466
#> [4,]  0.60749525  0.18846169
#> [5,]  0.68286897  0.23165690
#> [6,]  0.11549970 -0.08978224
#> 
#> , , 66
#> 
#>            [,1]        [,2]
#> [1,] -0.4990237  0.76247978
#> [2,]  0.1207963 -0.30557898
#> [3,] -0.3869024 -0.38651721
#> [4,]  0.4279868  0.01585734
#> [5,]  0.6085877  0.37033408
#> [6,] -0.1820396 -0.19609114
#> 
#> , , 67
#> 
#>             [,1]       [,2]
#> [1,] -0.46671651  0.6776423
#> [2,] -0.16820992 -0.3985492
#> [3,] -0.12284495 -0.5503932
#> [4,]  0.53917593  0.2012691
#> [5,]  0.66927232  0.1061528
#> [6,] -0.01241451 -0.1650730
#> 
#> , , 68
#> 
#>             [,1]        [,2]
#> [1,] -0.33303927  0.53010106
#> [2,]  0.21815693 -0.58793254
#> [3,] -0.25950373 -0.35161527
#> [4,]  0.47569809  0.02102031
#> [5,]  0.74007895  0.28175206
#> [6,]  0.01202844 -0.41215147
#> 
#> , , 69
#> 
#>             [,1]        [,2]
#> [1,] -0.57061283  0.76239974
#> [2,] -0.03620974 -0.17070301
#> [3,] -0.39140378 -0.44251899
#> [4,]  0.30939547 -0.06718678
#> [5,]  0.62869786  0.38481664
#> [6,] -0.17001876 -0.20294375
#> 
#> , , 70
#> 
#>             [,1]         [,2]
#> [1,] -0.43736328  0.720596886
#> [2,] -0.08562557 -0.311643208
#> [3,] -0.12733300 -0.553740589
#> [4,]  0.42010688  0.003594633
#> [5,]  0.77984848  0.277324599
#> [6,]  0.02268221  0.008255042
#> 
#> , , 71
#> 
#>             [,1]        [,2]
#> [1,] -0.43748412  0.40316546
#> [2,]  0.09331317 -0.15833915
#> [3,] -0.46982355 -0.63167826
#> [4,]  0.44078523  0.16131785
#> [5,]  0.58206853 -0.07583171
#> [6,]  0.21464097 -0.61773377
#> 
#> , , 72
#> 
#>              [,1]        [,2]
#> [1,] -0.450625457  0.75539472
#> [2,] -0.001821622 -0.29819053
#> [3,] -0.234183385 -0.40169724
#> [4,]  0.486097148  0.09062232
#> [5,]  0.710818414  0.29331419
#> [6,]  0.023200572 -0.29129870
#> 
#> , , 73
#> 
#>             [,1]       [,2]
#> [1,] -0.54183305  0.5732584
#> [2,] -0.34052897 -0.2006677
#> [3,] -0.57587282 -0.2460799
#> [4,]  0.27732879 -0.1108035
#> [5,]  0.41832881  0.4346809
#> [6,]  0.08316971 -0.6077228
#> 
#> , , 74
#> 
#>             [,1]        [,2]
#> [1,] -0.36910402  0.65330596
#> [2,]  0.08442342  0.03253359
#> [3,] -0.13853124 -0.57648684
#> [4,]  0.68610834 -0.09303102
#> [5,]  0.59952912  0.31986410
#> [6,] -0.08523017 -0.35892618
#> 
#> , , 75
#> 
#>             [,1]       [,2]
#> [1,] -0.47533436  0.7532152
#> [2,] -0.05009894 -0.2771332
#> [3,] -0.26216428 -0.4244620
#> [4,]  0.44161871  0.1195530
#> [5,]  0.71196112  0.2656311
#> [6,]  0.03002514 -0.3014021
#> 
#> , , 76
#> 
#>             [,1]        [,2]
#> [1,] -0.49562892  0.78911778
#> [2,] -0.02690285 -0.35633742
#> [3,] -0.17576825 -0.26645680
#> [4,]  0.30750043  0.06613851
#> [5,]  0.78214978  0.36993107
#> [6,] -0.12813639 -0.19517748
#> 
#> , , 77
#> 
#>              [,1]        [,2]
#> [1,] -0.471053902  0.71023229
#> [2,] -0.004514309  0.05162701
#> [3,] -0.431312105 -0.41334891
#> [4,]  0.564135352  0.12602822
#> [5,]  0.517075787  0.24238824
#> [6,]  0.080259863 -0.49740553
#> 
#> , , 78
#> 
#>              [,1]       [,2]
#> [1,] -0.346363556  0.6273173
#> [2,]  0.001266242 -0.3651047
#> [3,] -0.292055473 -0.2308027
#> [4,]  0.575962378  0.2413822
#> [5,]  0.641060699  0.2171957
#> [6,]  0.228128923 -0.5607694
#> 
#> , , 79
#> 
#>             [,1]       [,2]
#> [1,] -0.45585404  0.7630347
#> [2,] -0.04806865 -0.3104394
#> [3,] -0.16803354 -0.3868130
#> [4,]  0.47816226  0.1122237
#> [5,]  0.72987146  0.2869717
#> [6,] -0.01731258 -0.2771897
#> 
#> , , 80
#> 
#>             [,1]       [,2]
#> [1,] -0.35483372  0.5880354
#> [2,] -0.06349432 -0.3473614
#> [3,]  0.09745829 -0.4175022
#> [4,]  0.18012361 -0.2656999
#> [5,]  0.90167526  0.3595573
#> [6,]  0.12288451 -0.3992099
#> 
#> , , 81
#> 
#>             [,1]       [,2]
#> [1,] -0.44462944  0.7639482
#> [2,]  0.10820144 -0.2471366
#> [3,] -0.23795622 -0.4058183
#> [4,]  0.53510519  0.2171044
#> [5,]  0.66610833  0.2004304
#> [6,] -0.06273817 -0.3214210
#> 
#> , , 82
#> 
#>            [,1]       [,2]
#> [1,] -0.4595728  0.7591956
#> [2,] -0.3182369 -0.3330911
#> [3,] -0.2102575 -0.1141027
#> [4,]  0.6111835 -0.0734729
#> [5,]  0.5160302  0.5262727
#> [6,]  0.0589698 -0.1314982
#> 
#> , , 83
#> 
#>             [,1]        [,2]
#> [1,] -0.30226619  0.75581584
#> [2,]  0.28620660 -0.19699823
#> [3,] -0.22206541 -0.39975400
#> [4,]  0.61218607 -0.06641258
#> [5,]  0.63102455  0.40190959
#> [6,]  0.06666423 -0.25335529
#> 
#> , , 84
#> 
#>             [,1]       [,2]
#> [1,] -0.51738266  0.7446829
#> [2,] -0.39734102 -0.2384961
#> [3,] -0.31800515 -0.3690741
#> [4,]  0.48935804 -0.1108552
#> [5,]  0.48327415  0.4747592
#> [6,]  0.01681732 -0.1211041
#> 
#> , , 85
#> 
#>             [,1]        [,2]
#> [1,] -0.40109208  0.80848063
#> [2,]  0.14250774 -0.22542722
#> [3,] -0.06961345 -0.24327571
#> [4,]  0.47648014 -0.02346641
#> [5,]  0.75829214  0.42988411
#> [6,] -0.10922622 -0.22584850
#> 
#> , , 86
#> 
#>            [,1]        [,2]
#> [1,] -0.5200256  0.58522167
#> [2,] -0.1959405 -0.10527992
#> [3,] -0.1998811 -0.58551439
#> [4,]  0.3783504 -0.32087080
#> [5,]  0.6833412  0.44239839
#> [6,] -0.2027907  0.07021554
#> 
#> , , 87
#> 
#>             [,1]        [,2]
#> [1,] -0.18629602  0.69954973
#> [2,]  0.15717015 -0.24249083
#> [3,]  0.06650179 -0.63121323
#> [4,]  0.83797236  0.17198976
#> [5,]  0.48227777  0.14209000
#> [6,] -0.03713918  0.06023441
#> 
#> , , 88
#> 
#>              [,1]        [,2]
#> [1,] -0.386390050  0.62142898
#> [2,]  0.007636019 -0.20336435
#> [3,] -0.262744354 -0.65321710
#> [4,]  0.465398959  0.04650678
#> [5,]  0.723300177  0.16358951
#> [6,]  0.204573918 -0.34183618
#> 
#> , , 89
#> 
#>            [,1]        [,2]
#> [1,] -0.3110467  0.62802494
#> [2,] -0.2377388 -0.24436564
#> [3,] -0.1709630 -0.60523917
#> [4,]  0.5752807  0.21500118
#> [5,]  0.5817283  0.08101769
#> [6,]  0.3848976 -0.35604253
#> 
#> , , 90
#> 
#>             [,1]       [,2]
#> [1,] -0.23938259  0.8004156
#> [2,]  0.11924333 -0.1227615
#> [3,] -0.08771717 -0.1337432
#> [4,]  0.30279513  0.3364886
#> [5,]  0.85166802  0.2544593
#> [6,]  0.32211705 -0.3852312
#> 
#> , , 91
#> 
#>             [,1]       [,2]
#> [1,] -0.49835322  0.6795232
#> [2,] -0.02267085 -0.3615273
#> [3,] -0.25107640 -0.4113903
#> [4,]  0.46478990  0.1022530
#> [5,]  0.68430775  0.2246384
#> [6,] -0.06151418 -0.4211724
#> 
#> , , 92
#> 
#>              [,1]       [,2]
#> [1,] -0.458476361  0.7557312
#> [2,]  0.006697462 -0.2987400
#> [3,] -0.241758063 -0.4208021
#> [4,]  0.470873355  0.1155253
#> [5,]  0.713850396  0.2701861
#> [6,]  0.001872331 -0.2760502
#> 
#> , , 93
#> 
#>              [,1]        [,2]
#> [1,] -0.462197397  0.72362406
#> [2,] -0.001603021 -0.21643631
#> [3,] -0.323856633 -0.36500302
#> [4,]  0.601982298  0.07847916
#> [5,]  0.564887456  0.30071532
#> [6,]  0.002711698 -0.44688662
#> 
#> , , 94
#> 
#>             [,1]        [,2]
#> [1,] -0.28193763  0.72433719
#> [2,]  0.36672940 -0.25789435
#> [3,] -0.20756453 -0.55113801
#> [4,]  0.54447256 -0.02892423
#> [5,]  0.66764654  0.30417605
#> [6,]  0.02711849 -0.10822829
#> 
#> , , 95
#> 
#>             [,1]         [,2]
#> [1,] -0.35010275  0.845312977
#> [2,] -0.09223511 -0.133094776
#> [3,] -0.29641027 -0.170363552
#> [4,]  0.83918224  0.122946392
#> [5,]  0.27467778  0.472844653
#> [6,] -0.03724229 -0.003182789
#> 
#> , , 96
#> 
#>             [,1]       [,2]
#> [1,] -0.46146832  0.7646133
#> [2,] -0.11158603 -0.3114346
#> [3,] -0.18879277 -0.3559174
#> [4,]  0.39602003  0.1818786
#> [5,]  0.76110455  0.2563592
#> [6,]  0.05329941 -0.3047918
#> 
#> , , 97
#> 
#>             [,1]        [,2]
#> [1,] -0.44120080  0.76955966
#> [2,]  0.04110087 -0.26163664
#> [3,] -0.23845334 -0.31850760
#> [4,]  0.41662743  0.06349256
#> [5,]  0.75593383  0.34671903
#> [6,]  0.04216865 -0.33709304
#> 
#> , , 98
#> 
#>             [,1]       [,2]
#> [1,] -0.39962640  0.7600114
#> [2,]  0.06304027 -0.4376670
#> [3,] -0.15325620 -0.2909825
#> [4,]  0.48750611  0.2756136
#> [5,]  0.75801707  0.2062942
#> [6,]  0.02418930 -0.1662508
#> 
#> , , 99
#> 
#>              [,1]       [,2]
#> [1,] -0.449800544  0.7582889
#> [2,]  0.003194186 -0.3147873
#> [3,] -0.238600873 -0.4010113
#> [4,]  0.462613613  0.1000615
#> [5,]  0.725002506  0.2882836
#> [6,]  0.033149715 -0.2682855
#> 
#> , , 100
#> 
#>               [,1]       [,2]
#> [1,] -0.2865696073  0.4689886
#> [2,] -0.0005118151 -0.1957087
#> [3,]  0.0273894240 -0.3090242
#> [4,]  0.7214903473 -0.2347160
#> [5,]  0.6262405079  0.4309201
#> [6,] -0.0663469869 -0.6367638
#> 
#> attr(,"class")
#> [1] "history_array"