Risk Dataset

Risk Dataset

Usage

risk

Format

A 563 x 6 array of Likert scale data

Source

https://statistik.boku.ac.at/nachlass_leisch/MSA/

Details

Adult Australian residents that have undertaken at least one holiday in the last year, which involved staying away from home for at least four nights, were asked what risks they have taken in the past. The questions were about recreational, health, career, financial, safety and social risk. The response options were on the following: never (1), rarely (2), quite often (3), often (4) or very often (5)

Examples

data(risk)
head(risk)
#>   Recreational Health Career Financial Safety Social
#> 1            3      2      1         2      2      4
#> 2            1      1      1         1      1      1
#> 3            2      2      1         1      2      3
#> 4            1      1      1         1      1      1
#> 5            5      4      1         3      5      5
#> 6            5      2      5         3      2      3