Austrian Vacation Activities

The Austrian Vacation Activities dataset comprises responses from 2,961 adult tourists who spent their holiday in Austria during the 1997/98 season. The responses are coded in binary with 0 indicating that the tourist didn't partake in the activity and 1 indicating they did.

Usage

winterActiv

Format

A 2961 x 27 array of binary responses

Examples

head(winterActiv)
#>      alpine.skiing cross.country.skiing snowboarding carving ski.touring
#> [1,]             1                    0            0       0           0
#> [2,]             1                    0            0       0           0
#> [3,]             0                    0            0       0           0
#> [4,]             1                    0            0       0           0
#> [5,]             1                    0            0       0           0
#> [6,]             1                    0            0       0           0
#>      ice.skating sleigh.riding tennis horseback.riding going.to.a.spa
#> [1,]           0             0      0                0              1
#> [2,]           1             1      0                0              1
#> [3,]           0             0      0                0              1
#> [4,]           0             0      0                0              1
#> [5,]           0             0      0                0              1
#> [6,]           0             0      0                0              0
#>      using.health.facilities hiking going.for.walks organized.excursions
#> [1,]                       0      0               0                    0
#> [2,]                       0      1               1                    0
#> [3,]                       0      0               0                    0
#> [4,]                       0      0               1                    0
#> [5,]                       0      1               1                    0
#> [6,]                       0      1               1                    0
#>      excursions relaxing going.out.in.the.evening going.to.discos.bars shopping
#> [1,]          0        1                        1                    0        0
#> [2,]          0        1                        0                    0        0
#> [3,]          0        1                        1                    0        1
#> [4,]          0        1                        1                    0        1
#> [5,]          1        0                        1                    0        1
#> [6,]          0        0                        1                    1        1
#>      sight.seeing museums theater.opera heurigen concerts tyrolean.evenings
#> [1,]            0       0             0        0        0                 0
#> [2,]            1       1             0        0        0                 0
#> [3,]            0       0             0        0        0                 0
#> [4,]            0       0             0        0        0                 0
#> [5,]            1       1             0        0        0                 0
#> [6,]            0       0             0        0        0                 0
#>      local.events pool.sauna
#> [1,]            0          1
#> [2,]            0          1
#> [3,]            0          1
#> [4,]            0          0
#> [5,]            0          1
#> [6,]            0          1