Get Guided Tour-Holes-Better History

Returns a guided tour with the holes index and the 'search_better' argument to the python' backend. This guided tour is generated with the 'tourr' functions 'save_history' and 'guided_tour'.

Usage

get_guided_holes_better_history(data, dimension)

Arguments

data

the dataset to calculate the projections with.

dimension

1 for a 1d tour or 2 for a 2d tour

Value

history object containing the projections of the requested tour

Examples

data("flea", package = "tourr")
flea <- flea[-7]
get_guided_holes_better_history(flea, 2)
#> Converting input data to the required matrix format.
#> Target: 1.090, try: 1
#> Target: 1.113, try: 2
#> Target: 1.154, try: 12
#> No better bases found after 25 tries.  Giving up.
#> Final projection: 
#> -0.302  -0.474  
#> -0.485   0.050  
#> -0.739  -0.062  
#> -0.293   0.751  
#>  0.174   0.045  
#>  0.104   0.451  
#> , , 1
#> 
#>              [,1]       [,2]
#> [1,] -0.450205705 -0.7646855
#> [2,]  0.082101158 -0.1961486
#> [3,] -0.783739912  0.3638870
#> [4,] -0.001791534 -0.1625485
#> [5,]  0.199869917 -0.4494856
#> [6,]  0.369289562  0.1261300
#> 
#> , , 2
#> 
#>            [,1]        [,2]
#> [1,]  0.1708642 -0.42883081
#> [2,] -0.4030014 -0.36887408
#> [3,] -0.6162850  0.22248962
#> [4,] -0.5540070  0.09216859
#> [5,]  0.0411629 -0.67616445
#> [6,]  0.3463667  0.40600616
#> 
#> , , 3
#> 
#>             [,1]       [,2]
#> [1,] -0.59833206 -0.1062776
#> [2,] -0.48813654 -0.2583171
#> [3,] -0.62166077  0.3121632
#> [4,] -0.07583558  0.6480557
#> [5,] -0.01042725 -0.4411375
#> [6,]  0.10676886  0.4582063
#> 
#> , , 4
#> 
#>            [,1]        [,2]
#> [1,] -0.4844725 -0.28532361
#> [2,] -0.4096992  0.26410262
#> [3,] -0.6871202  0.28011366
#> [4,]  0.0795257  0.80179860
#> [5,]  0.1752625 -0.03865497
#> [6,]  0.2970817  0.35496673
#> 
#> , , 5
#> 
#>            [,1]        [,2]
#> [1,] -0.4844725 -0.28532361
#> [2,] -0.4096992  0.26410262
#> [3,] -0.6871202  0.28011366
#> [4,]  0.0795257  0.80179860
#> [5,]  0.1752625 -0.03865497
#> [6,]  0.2970817  0.35496673
#> 
#> attr(,"class")
#> [1] "history_array"