Get Guided Tour-Holes History

Returns a guided tour with the holes index to the python' backend. This guided tour is generated with the 'tourr' functions 'save_history' and 'guided_tour'.

Usage

get_guided_holes_history(data, dimension)

Arguments

data

the dataset to calculate the projections with.

dimension

1 for a 1d tour or 2 for a 2d tour

Value

history object containing the projections of the requested tour

Examples

data("flea", package = "tourr")
flea <- flea[-7]
get_guided_holes_history(flea, 2)
#> Converting input data to the required matrix format.
#> Target: 0.784, 25.2% better 
#> Target: 0.849, 8.3% better 
#> Target: 1.024, 20.7% better 
#> Target: 1.200, 17.2% better 
#> Target: 1.219, 1.5% better 
#> Target: 1.237, 1.5% better 
#> Target: 1.250, 1.0% better 
#> Target: 1.257, 0.5% better 
#> Target: 1.261, 0.3% better 
#> Target: 1.265, 0.3% better 
#> Target: 1.267, 0.2% better 
#> Target: 1.274, 0.5% better 
#> Target: 1.276, 0.2% better 
#> No better bases found after 25 tries.  Giving up.
#> Final projection: 
#>  0.503   0.523  
#> -0.130   0.649  
#> -0.263   0.386  
#> -0.485   0.296  
#>  0.376   0.256  
#> -0.533   0.056  
#> , , 1
#> 
#>            [,1]          [,2]
#> [1,] -0.1233320  0.1853455489
#> [2,]  0.3071400  0.2431149849
#> [3,]  0.3525569  0.6302843435
#> [4,]  0.4220958 -0.7091189517
#> [5,]  0.7011546  0.0005601995
#> [6,] -0.3104436 -0.0802108252
#> 
#> , , 2
#> 
#>              [,1]       [,2]
#> [1,]  0.440517281  0.2992482
#> [2,] -0.005600709  0.2150524
#> [3,]  0.322209912  0.6437882
#> [4,]  0.410958136 -0.5874893
#> [5,]  0.729379886 -0.1462403
#> [6,] -0.034818410 -0.2884611
#> 
#> , , 3
#> 
#>             [,1]       [,2]
#> [1,]  0.21825588  0.6685129
#> [2,]  0.20295547  0.1466540
#> [3,]  0.07922971  0.4383378
#> [4,]  0.14587391 -0.5276461
#> [5,]  0.85002081 -0.2146886
#> [6,] -0.40134964 -0.1222352
#> 
#> , , 4
#> 
#>             [,1]         [,2]
#> [1,]  0.24178180  0.481772746
#> [2,]  0.04833947  0.623085158
#> [3,] -0.23101540  0.380715513
#> [4,] -0.08701199 -0.457709374
#> [5,]  0.63472013 -0.158766084
#> [6,] -0.68948967 -0.003326348
#> 
#> , , 5
#> 
#>             [,1]        [,2]
#> [1,]  0.18488926  0.34751666
#> [2,]  0.02871987  0.75364527
#> [3,] -0.27938932  0.53223528
#> [4,] -0.38292567  0.10592191
#> [5,]  0.44543750  0.10344684
#> [6,] -0.73612914 -0.07781961
#> 
#> , , 6
#> 
#>             [,1]        [,2]
#> [1,]  0.17404499  0.42685633
#> [2,]  0.03543175  0.66828901
#> [3,] -0.21305680  0.54129202
#> [4,] -0.57018154  0.19176331
#> [5,]  0.43742279  0.19221583
#> [6,] -0.63766296 -0.06683062
#> 
#> , , 7
#> 
#>             [,1]       [,2]
#> [1,]  0.30676624  0.4103858
#> [2,] -0.05462251  0.7252835
#> [3,] -0.11857470  0.4855894
#> [4,] -0.66244560  0.2062514
#> [5,]  0.31219928  0.1070798
#> [6,] -0.59375780 -0.1254771
#> 
#> , , 8
#> 
#>             [,1]        [,2]
#> [1,]  0.42812018  0.47160224
#> [2,] -0.16632692  0.69754191
#> [3,] -0.08218638  0.48243430
#> [4,] -0.60394796  0.20513492
#> [5,]  0.34636805  0.02900518
#> [6,] -0.54549968 -0.12394405
#> 
#> , , 9
#> 
#>            [,1]       [,2]
#> [1,]  0.4313556  0.4998210
#> [2,] -0.1177439  0.6069341
#> [3,] -0.1519289  0.5249215
#> [4,] -0.5922161  0.2843792
#> [5,]  0.4240473  0.1012058
#> [6,] -0.4964377 -0.1230988
#> 
#> , , 10
#> 
#>            [,1]       [,2]
#> [1,]  0.4892279  0.4662348
#> [2,] -0.1207205  0.6899447
#> [3,] -0.1768715  0.4599753
#> [4,] -0.6055305  0.2623299
#> [5,]  0.3803215  0.1103355
#> [6,] -0.4510957 -0.1184620
#> 
#> , , 11
#> 
#>            [,1]       [,2]
#> [1,]  0.5098975  0.4056433
#> [2,] -0.1330552  0.6359793
#> [3,] -0.1956134  0.5054328
#> [4,] -0.5808030  0.3075875
#> [5,]  0.3978584  0.2567402
#> [6,] -0.4340655 -0.1224579
#> 
#> , , 12
#> 
#>            [,1]       [,2]
#> [1,]  0.4799100  0.4393549
#> [2,] -0.1406906  0.6156144
#> [3,] -0.2107614  0.5099030
#> [4,] -0.5358874  0.3210051
#> [5,]  0.4078653  0.2177762
#> [6,] -0.5019389 -0.1323417
#> 
#> , , 13
#> 
#>            [,1]         [,2]
#> [1,]  0.5428138  0.513680503
#> [2,] -0.1387059  0.670557343
#> [3,] -0.2617445  0.380298077
#> [4,] -0.5154364  0.312823219
#> [5,]  0.3433877  0.209672993
#> [6,] -0.4837498 -0.006118254
#> 
#> , , 14
#> 
#>            [,1]       [,2]
#> [1,]  0.5026333 0.52314154
#> [2,] -0.1298222 0.64916552
#> [3,] -0.2629996 0.38615590
#> [4,] -0.4854807 0.29563372
#> [5,]  0.3758889 0.25554063
#> [6,] -0.5332477 0.05559109
#> 
#> , , 15
#> 
#>            [,1]       [,2]
#> [1,]  0.5026333 0.52314154
#> [2,] -0.1298222 0.64916552
#> [3,] -0.2629996 0.38615590
#> [4,] -0.4854807 0.29563372
#> [5,]  0.3758889 0.25554063
#> [6,] -0.5332477 0.05559109
#> 
#> attr(,"class")
#> [1] "history_array"