Get Local Tour History

Returns a local tour based on currently displayed projection(s) to the 'python' backend. This local tour is generated with the 'tourr' functions 'save_history' and 'local_tour'.

Usage

get_local_history(data, starting_projection)

Arguments

data

the dataset to calculate the projections with. In practice only the two first rows of the dataset are provided as the actual data is not needed.

starting_projection

the initial projection one wants to initiate the local tour from

Value

history object containing the projections of the requested tour

Examples

library(tourr)
data("flea", package = "tourr")
flea <- flea[-7]
prj <- tourr::basis_random(ncol(flea), 2)
get_local_history(flea, prj)
#> Converting input data to the required matrix format.
#> , , 1
#> 
#>             [,1]        [,2]
#> [1,]  0.68506812  0.14606613
#> [2,] -0.35735041  0.54501451
#> [3,] -0.02488716  0.64468939
#> [4,]  0.63162967  0.17067267
#> [5,] -0.05671157  0.06562487
#> [6,] -0.01381065 -0.48224857
#> 
#> , , 2
#> 
#>              [,1]       [,2]
#> [1,]  0.547723209  0.1165561
#> [2,] -0.246684096  0.5427528
#> [3,] -0.112236186  0.5593832
#> [4,]  0.783177419  0.1927087
#> [5,] -0.114480864  0.1142355
#> [6,]  0.008748953 -0.5733568
#> 
#> , , 3
#> 
#>              [,1]       [,2]
#> [1,]  0.617965545  0.1142067
#> [2,] -0.451431780  0.5324056
#> [3,]  0.009695745  0.6216592
#> [4,]  0.631500033  0.1889211
#> [5,] -0.084510000  0.0440021
#> [6,] -0.091102658 -0.5285958
#> 
#> , , 4
#> 
#>             [,1]        [,2]
#> [1,]  0.72409446 -0.04349169
#> [2,] -0.27874687  0.54675919
#> [3,] -0.31610073  0.32038151
#> [4,]  0.50467275  0.26708153
#> [5,]  0.06449068  0.16517964
#> [6,] -0.19802546 -0.70562149
#> 
#> , , 5
#> 
#>             [,1]       [,2]
#> [1,]  0.82131689  0.1874609
#> [2,] -0.31426822  0.6660041
#> [3,] -0.02389185  0.5638698
#> [4,]  0.47215004  0.1826358
#> [5,]  0.02495951  0.1078743
#> [6,]  0.05054299 -0.3979386
#> 
#> , , 6
#> 
#>             [,1]        [,2]
#> [1,]  0.65923951 -0.01640786
#> [2,] -0.47171358  0.52370658
#> [3,]  0.22180431  0.64759990
#> [4,]  0.52418974  0.33340825
#> [5,]  0.01594489 -0.08381781
#> [6,]  0.13661366 -0.43346290
#> 
#> , , 7
#> 
#>              [,1]        [,2]
#> [1,]  0.553527838  0.19134322
#> [2,] -0.531325585  0.56220737
#> [3,] -0.002076814  0.38822908
#> [4,]  0.603540320  0.07567385
#> [5,] -0.039659913 -0.28195273
#> [6,] -0.213217998 -0.64137739
#> 
#> , , 8
#> 
#>              [,1]        [,2]
#> [1,]  0.643054908  0.17162274
#> [2,] -0.431103310  0.61131406
#> [3,]  0.071002132  0.67359967
#> [4,]  0.628351975  0.18213738
#> [5,] -0.001729931  0.05863005
#> [6,]  0.027564843 -0.32633223
#> 
#> , , 9
#> 
#>              [,1]        [,2]
#> [1,]  0.679857622  0.22182043
#> [2,] -0.426824871  0.55495207
#> [3,]  0.010086959  0.59660037
#> [4,]  0.594668440  0.09692909
#> [5,]  0.007730759  0.19431559
#> [6,] -0.042685802 -0.48963058
#> 
#> , , 10
#> 
#>              [,1]        [,2]
#> [1,]  0.732034400  0.09689606
#> [2,] -0.185027181  0.58641433
#> [3,] -0.168185924  0.73929579
#> [4,]  0.633431683  0.26017225
#> [5,]  0.018924409 -0.13200555
#> [6,] -0.003200684  0.12270310
#> 
#> , , 11
#> 
#>               [,1]        [,2]
#> [1,]  0.7522188756  0.11524259
#> [2,] -0.3108464302  0.53085074
#> [3,] -0.0505524671  0.62187537
#> [4,]  0.5754866924  0.19410892
#> [5,] -0.0616502317  0.03021272
#> [6,]  0.0001545789 -0.52876893
#> 
#> , , 12
#> 
#>             [,1]       [,2]
#> [1,]  0.86561091  0.1588749
#> [2,] -0.41281303  0.5338266
#> [3,]  0.08188255  0.4418375
#> [4,]  0.26682492  0.2020635
#> [5,] -0.03905244  0.5149894
#> [6,] -0.02962732 -0.4341933
#> 
#> , , 13
#> 
#>             [,1]        [,2]
#> [1,]  0.25421476  0.05007291
#> [2,] -0.62790756  0.40306452
#> [3,]  0.24477555  0.48120429
#> [4,]  0.69020489  0.12418782
#> [5,] -0.06272588 -0.24497883
#> [6,] -0.02957276 -0.72666147
#> 
#> , , 14
#> 
#>             [,1]       [,2]
#> [1,]  0.61325333  0.1335098
#> [2,] -0.45251841  0.4172292
#> [3,]  0.06923732  0.7850229
#> [4,]  0.64169682  0.1005087
#> [5,] -0.04802253  0.1062704
#> [6,]  0.01651261 -0.4128420
#> 
#> , , 15
#> 
#>             [,1]        [,2]
#> [1,]  0.63298971  0.08191212
#> [2,] -0.39655631  0.46103394
#> [3,]  0.06399451  0.70066624
#> [4,]  0.65684046  0.18763290
#> [5,]  0.01499345  0.12051027
#> [6,]  0.07942059 -0.48997562
#> 
#> , , 16
#> 
#>             [,1]       [,2]
#> [1,]  0.72357505  0.2257763
#> [2,] -0.22763114  0.5867403
#> [3,] -0.06179108  0.6970249
#> [4,]  0.62422633  0.1011828
#> [5,]  0.13412712 -0.1046814
#> [6,]  0.11470166 -0.3126035
#> 
#> , , 17
#> 
#>            [,1]        [,2]
#> [1,]  0.5844906  0.06275643
#> [2,] -0.2729219  0.71945782
#> [3,]  0.1527468  0.44820279
#> [4,]  0.6686458  0.33926198
#> [5,] -0.1065880  0.11023376
#> [6,]  0.3195380 -0.38769339
#> 
#> , , 18
#> 
#>            [,1]        [,2]
#> [1,]  0.9166192  0.01589553
#> [2,] -0.1661271  0.28836533
#> [3,] -0.1131105  0.48233913
#> [4,]  0.2588781  0.20181008
#> [5,] -0.2200651  0.06685315
#> [6,] -0.0630126 -0.79921529
#> 
#> , , 19
#> 
#>             [,1]        [,2]
#> [1,]  0.65201501 -0.16937752
#> [2,] -0.07436857  0.44222238
#> [3,] -0.20859880  0.67999420
#> [4,]  0.65966250  0.43894244
#> [5,] -0.29817808 -0.03404425
#> [6,]  0.04204155 -0.34572971
#> 
#> , , 20
#> 
#>             [,1]       [,2]
#> [1,]  0.63597136  0.1190331
#> [2,] -0.34722920  0.4949934
#> [3,] -0.08457172  0.6666111
#> [4,]  0.67992938  0.1750102
#> [5,] -0.01914444  0.1234833
#> [6,] -0.07175971 -0.5005653
#> 
#> , , 21
#> 
#>             [,1]        [,2]
#> [1,]  0.74861452  0.31999713
#> [2,] -0.37625739  0.57558551
#> [3,]  0.00785142  0.55776297
#> [4,]  0.53633155 -0.02714898
#> [5,] -0.10021880  0.04859847
#> [6,]  0.01580181 -0.50210034
#> 
#> , , 22
#> 
#>             [,1]        [,2]
#> [1,]  0.64930606  0.22396995
#> [2,] -0.24351927  0.49942687
#> [3,] -0.15171901  0.66003363
#> [4,]  0.70158956  0.07285183
#> [5,]  0.04178573  0.04633576
#> [6,] -0.04590639 -0.50725879
#> 
#> , , 23
#> 
#>              [,1]        [,2]
#> [1,]  0.800716790  0.08066767
#> [2,] -0.334437902  0.40924988
#> [3,] -0.128991546  0.70563668
#> [4,]  0.454423704  0.18757304
#> [5,]  0.009238079  0.22368506
#> [6,] -0.154203914 -0.49281385
#> 
#> , , 24
#> 
#>             [,1]         [,2]
#> [1,]  0.81724430 -0.007997648
#> [2,] -0.37783454  0.561580693
#> [3,] -0.02942362  0.512520837
#> [4,]  0.41156736  0.395092204
#> [5,]  0.03385825  0.106583073
#> [6,] -0.13398878 -0.504408320
#> 
#> , , 25
#> 
#>             [,1]        [,2]
#> [1,]  0.71936988  0.23985573
#> [2,] -0.40607463  0.55599345
#> [3,] -0.09619077  0.66314624
#> [4,]  0.52723410  0.08055518
#> [5,]  0.04263982  0.16727760
#> [6,] -0.16900816 -0.39888175
#> 
#> , , 26
#> 
#>            [,1]       [,2]
#> [1,]  0.4902216  0.2234137
#> [2,] -0.5249250  0.6135202
#> [3,]  0.2739075  0.5236675
#> [4,]  0.5442118  0.2673198
#> [5,] -0.3245382  0.3285560
#> [6,] -0.0872909 -0.3464719
#> 
#> , , 27
#> 
#>             [,1]       [,2]
#> [1,]  0.68592829  0.1388297
#> [2,] -0.37210176  0.5478930
#> [3,] -0.01561471  0.6578962
#> [4,]  0.61191751  0.1909112
#> [5,] -0.12179548  0.1245017
#> [6,] -0.03900842 -0.4424528
#> 
#> , , 28
#> 
#>             [,1]        [,2]
#> [1,]  0.70622433  0.17222755
#> [2,] -0.36227047  0.58089722
#> [3,] -0.03036144  0.63512876
#> [4,]  0.60221256  0.16622439
#> [5,] -0.07276110  0.09333716
#> [6,] -0.03363539 -0.43950561
#> 
#> , , 29
#> 
#>             [,1]       [,2]
#> [1,]  0.67847017  0.1794362
#> [2,] -0.33348428  0.7174678
#> [3,] -0.07503510  0.2891493
#> [4,]  0.62351438  0.2141579
#> [5,]  0.17929779 -0.1048491
#> [6,] -0.04379855 -0.5590871
#> 
#> , , 30
#> 
#>              [,1]        [,2]
#> [1,]  0.629173493  0.14859017
#> [2,] -0.339273077  0.56929194
#> [3,] -0.001116250  0.64904602
#> [4,]  0.697176288  0.17846239
#> [5,]  0.006757819  0.02670832
#> [6,]  0.054155375 -0.44721892
#> 
#> , , 31
#> 
#>             [,1]       [,2]
#> [1,]  0.80621139  0.1763643
#> [2,] -0.21157989  0.4614669
#> [3,] -0.07584483  0.6329398
#> [4,]  0.53710024  0.1134258
#> [5,] -0.07541105  0.2486090
#> [6,]  0.07308369 -0.5297728
#> 
#> , , 32
#> 
#>             [,1]        [,2]
#> [1,]  0.68080839  0.14573680
#> [2,] -0.35624737  0.55252560
#> [3,] -0.02786119  0.63677395
#> [4,]  0.63483103  0.17547748
#> [5,] -0.07374173  0.07083737
#> [6,] -0.01905832 -0.48185568
#> 
#> , , 33
#> 
#>             [,1]        [,2]
#> [1,]  0.62408382  0.15691257
#> [2,] -0.38669703  0.55705963
#> [3,]  0.01796910  0.63638916
#> [4,]  0.67834360  0.16679436
#> [5,] -0.01909703  0.06833849
#> [6,]  0.01213164 -0.47705487
#> 
#> , , 34
#> 
#>            [,1]        [,2]
#> [1,]  0.4751113  0.25495299
#> [2,] -0.4383196  0.45725966
#> [3,] -0.2046270  0.64603662
#> [4,]  0.5757155 -0.06992997
#> [5,] -0.2285095 -0.26462256
#> [6,] -0.3957373 -0.48335697
#> 
#> , , 35
#> 
#>            [,1]       [,2]
#> [1,]  0.3831885 -0.0741985
#> [2,] -0.4014303  0.3378776
#> [3,] -0.1638517  0.8587403
#> [4,]  0.6554757  0.2800335
#> [5,] -0.3276798 -0.1349024
#> [6,] -0.3579811 -0.2151299
#> 
#> , , 36
#> 
#>             [,1]       [,2]
#> [1,]  0.68326082  0.1466638
#> [2,] -0.35787050  0.5454583
#> [3,] -0.02505253  0.6415710
#> [4,]  0.63318803  0.1700005
#> [5,] -0.05754803  0.0648762
#> [6,] -0.01472686 -0.4860478
#> 
#> , , 37
#> 
#>            [,1]       [,2]
#> [1,]  0.4535707  0.2115452
#> [2,] -0.3722895  0.7437447
#> [3,]  0.1509838  0.5087012
#> [4,]  0.7122724  0.1612012
#> [5,] -0.3114863 -0.1361160
#> [6,]  0.1688855 -0.3143280
#> 
#> , , 38
#> 
#>              [,1]       [,2]
#> [1,]  0.681355132  0.1602479
#> [2,] -0.474466161  0.6023428
#> [3,]  0.124419983  0.5961481
#> [4,]  0.529465397  0.1915069
#> [5,] -0.121610865  0.0138959
#> [6,] -0.005822817 -0.4682341
#> 
#> , , 39
#> 
#>             [,1]       [,2]
#> [1,]  0.59791963  0.2314130
#> [2,] -0.29267382  0.5991539
#> [3,] -0.08355933  0.6383346
#> [4,]  0.69826316  0.1328223
#> [5,] -0.23971943  0.1213013
#> [6,] -0.06939128 -0.3842341
#> 
#> , , 40
#> 
#>             [,1]        [,2]
#> [1,]  0.66208574  0.14214787
#> [2,] -0.35268816  0.50076236
#> [3,] -0.02728624  0.66143303
#> [4,]  0.65875963  0.15227950
#> [5,] -0.04966714  0.08612667
#> [6,] -0.00882705 -0.51081361
#> 
#> , , 41
#> 
#>              [,1]       [,2]
#> [1,]  0.576742639 -0.1873809
#> [2,] -0.049817318  0.5066380
#> [3,] -0.377677798  0.5682910
#> [4,]  0.721015726  0.4689811
#> [5,] -0.009930667  0.3053151
#> [6,] -0.047784394 -0.2684979
#> 
#> , , 42
#> 
#>             [,1]       [,2]
#> [1,]  0.65559424  0.1537221
#> [2,] -0.32925904  0.5154077
#> [3,]  0.04978741  0.7098408
#> [4,]  0.61608324  0.2180051
#> [5,] -0.28156165  0.3418750
#> [6,]  0.02168815 -0.2060235
#> 
#> , , 43
#> 
#>             [,1]        [,2]
#> [1,]  0.74227551  0.09842365
#> [2,] -0.27420773  0.49577961
#> [3,] -0.04987821  0.70665489
#> [4,]  0.60527828  0.21375520
#> [5,] -0.04330426  0.22311769
#> [6,]  0.05578791 -0.38688687
#> 
#> , , 44
#> 
#>             [,1]        [,2]
#> [1,]  0.71106542  0.16680872
#> [2,] -0.28733305  0.42840431
#> [3,] -0.04888077  0.70090149
#> [4,]  0.63790879  0.09382202
#> [5,]  0.02351817  0.09629943
#> [6,]  0.04422242 -0.52849367
#> 
#> , , 45
#> 
#>             [,1]         [,2]
#> [1,]  0.66083190  0.148853841
#> [2,] -0.36924586  0.516645662
#> [3,]  0.00292591  0.682409696
#> [4,]  0.65242861  0.148150651
#> [5,] -0.03350724  0.005848299
#> [6,]  0.01281812 -0.472497601
#> 
#> , , 46
#> 
#>              [,1]        [,2]
#> [1,]  0.668813654  0.36319071
#> [2,] -0.384726018  0.61602173
#> [3,]  0.001650992  0.68658544
#> [4,]  0.608055671 -0.02863351
#> [5,]  0.184605795  0.07289646
#> [6,]  0.029333654 -0.10524450
#> 
#> , , 47
#> 
#>              [,1]        [,2]
#> [1,]  0.691200097  0.15402693
#> [2,] -0.343082269  0.53936620
#> [3,] -0.026402216  0.64927724
#> [4,]  0.634553800  0.15820211
#> [5,] -0.033868345  0.05364081
#> [6,]  0.005857894 -0.48568882
#> 
#> , , 48
#> 
#>              [,1]        [,2]
#> [1,]  0.612978347  0.12213787
#> [2,] -0.407700856  0.44716791
#> [3,]  0.030310274  0.71749094
#> [4,]  0.675921428  0.13681409
#> [5,]  0.006359782  0.06385705
#> [6,]  0.014443751 -0.49752802
#> 
#> , , 49
#> 
#>             [,1]        [,2]
#> [1,]  0.80538776  0.02799137
#> [2,] -0.07485849  0.53297591
#> [3,] -0.19675652  0.42104665
#> [4,]  0.54762244  0.28417232
#> [5,] -0.02859977  0.06690491
#> [6,]  0.07953202 -0.67278726
#> 
#> , , 50
#> 
#>              [,1]        [,2]
#> [1,]  0.797884548  0.43799400
#> [2,] -0.206500586  0.47233981
#> [3,]  0.009885535  0.30066352
#> [4,]  0.524899004 -0.21832101
#> [5,] -0.024640524 -0.02615194
#> [6,]  0.210983202 -0.66806424
#> 
#> , , 51
#> 
#>             [,1]        [,2]
#> [1,]  0.80247483  0.07331328
#> [2,] -0.41746672  0.39984507
#> [3,] -0.01801036  0.66149864
#> [4,]  0.41276638  0.15600980
#> [5,]  0.06659036  0.09879827
#> [6,] -0.08136923 -0.60255164
#> 
#> , , 52
#> 
#>              [,1]          [,2]
#> [1,]  0.702143498  0.1050835313
#> [2,] -0.366319234  0.5543290085
#> [3,]  0.003470497  0.5996130978
#> [4,]  0.609379352  0.2160611458
#> [5,] -0.037123829  0.0009295705
#> [6,]  0.008444446 -0.5248405965
#> 
#> , , 53
#> 
#>             [,1]        [,2]
#> [1,]  0.75576976  0.09041682
#> [2,] -0.28179806  0.48117874
#> [3,] -0.06287642  0.66448132
#> [4,]  0.58145193  0.20353009
#> [5,]  0.07396384  0.16776323
#> [6,]  0.04349120 -0.49918673
#> 
#> , , 54
#> 
#>              [,1]        [,2]
#> [1,]  0.657885026  0.06541068
#> [2,] -0.415057852  0.45350124
#> [3,]  0.004288308  0.83315844
#> [4,]  0.570797754  0.21805614
#> [5,] -0.243566286  0.01844170
#> [6,] -0.098799149 -0.21912668
#> 
#> , , 55
#> 
#>              [,1]        [,2]
#> [1,]  0.675307413  0.14785691
#> [2,] -0.353444399  0.53514935
#> [3,] -0.022253300  0.65438875
#> [4,]  0.644612865  0.16368035
#> [5,] -0.054765294  0.06619235
#> [6,] -0.004094246 -0.48203338
#> 
#> , , 56
#> 
#>             [,1]        [,2]
#> [1,]  0.61820944  0.15763785
#> [2,] -0.07478257  0.51790307
#> [3,] -0.16130829  0.71107034
#> [4,]  0.74795629  0.13989283
#> [5,] -0.09707866 -0.06937722
#> [6,]  0.13168681 -0.42062156
#> 
#> , , 57
#> 
#>             [,1]         [,2]
#> [1,]  0.62071449 -0.005727926
#> [2,] -0.51035481  0.698791657
#> [3,] -0.09802375  0.189899158
#> [4,]  0.44081068  0.325552666
#> [5,] -0.20210904 -0.285242040
#> [6,] -0.33087871 -0.536887474
#> 
#> , , 58
#> 
#>             [,1]        [,2]
#> [1,]  0.62627184  0.21492014
#> [2,] -0.40829094  0.52508148
#> [3,]  0.01645678  0.63216072
#> [4,]  0.66110291  0.09828339
#> [5,] -0.05415553  0.17657121
#> [6,] -0.02866001 -0.48747777
#> 
#> , , 59
#> 
#>            [,1]        [,2]
#> [1,]  0.5508008  0.18637762
#> [2,] -0.3506042  0.68391998
#> [3,]  0.0905306  0.56022514
#> [4,]  0.7284644  0.20637538
#> [5,]  0.1184425 -0.09816006
#> [6,]  0.1442583 -0.36254439
#> 
#> , , 60
#> 
#>              [,1]        [,2]
#> [1,]  0.469840962 -0.01019768
#> [2,] -0.388112077  0.53182883
#> [3,] -0.004756808  0.40582065
#> [4,]  0.706620834  0.29883398
#> [5,] -0.353194239  0.12209088
#> [6,] -0.067354918 -0.66944437
#> 
#> , , 61
#> 
#>              [,1]       [,2]
#> [1,]  0.631895847  0.1981577
#> [2,] -0.253601519  0.5754126
#> [3,] -0.110897045  0.6439707
#> [4,]  0.719205266  0.1383918
#> [5,] -0.082641386  0.1060191
#> [6,] -0.003152773 -0.4295850
#> 
#> , , 62
#> 
#>             [,1]        [,2]
#> [1,]  0.61924164  0.32983976
#> [2,] -0.19519767  0.69857985
#> [3,] -0.45969650  0.41204196
#> [4,]  0.51361978 -0.06288710
#> [5,]  0.00990846  0.03528519
#> [6,] -0.32126831 -0.47771699
#> 
#> , , 63
#> 
#>              [,1]        [,2]
#> [1,]  0.714265047  0.14755302
#> [2,] -0.403567819  0.54846753
#> [3,]  0.034785292  0.59891208
#> [4,]  0.562903850  0.16410685
#> [5,] -0.093957966  0.01518008
#> [6,] -0.007719968 -0.53995769
#> 
#> , , 64
#> 
#>              [,1]        [,2]
#> [1,]  0.686935019  0.14909714
#> [2,] -0.383819997  0.51232214
#> [3,]  0.003396486  0.64447145
#> [4,]  0.610445379  0.14407666
#> [5,] -0.089018408  0.04254596
#> [6,] -0.014937051 -0.52667294
#> 
#> , , 65
#> 
#>             [,1]        [,2]
#> [1,]  0.65125878 0.118969417
#> [2,] -0.46813906 0.625977744
#> [3,]  0.10562384 0.716673936
#> [4,]  0.53690104 0.280001496
#> [5,] -0.23127371 0.044307457
#> [6,] -0.06165356 0.003552806
#> 
#> , , 66
#> 
#>              [,1]        [,2]
#> [1,]  0.736001312  0.19167521
#> [2,] -0.374725222  0.56044040
#> [3,] -0.009015146  0.67369922
#> [4,]  0.563083006  0.12476636
#> [5,]  0.023106513 -0.05485494
#> [6,] -0.014332528 -0.42038175
#> 
#> , , 67
#> 
#>            [,1]       [,2]
#> [1,]  0.6464491  0.1820694
#> [2,] -0.1998051  0.6277471
#> [3,]  0.0426190  0.6245113
#> [4,]  0.6054503  0.2018913
#> [5,] -0.3744363  0.2312476
#> [6,]  0.1832825 -0.2975474
#> 
#> , , 68
#> 
#>              [,1]        [,2]
#> [1,]  0.691660371  0.12767668
#> [2,] -0.357662540  0.56163094
#> [3,] -0.003323167  0.63976017
#> [4,]  0.622579808  0.19847621
#> [5,] -0.077116178  0.04893083
#> [6,]  0.010948614 -0.46603566
#> 
#> , , 69
#> 
#>            [,1]        [,2]
#> [1,]  0.8695152  0.26296083
#> [2,] -0.1344576  0.53094727
#> [3,] -0.1282723  0.64082411
#> [4,]  0.4351479 -0.01643656
#> [5,]  0.0129613 -0.13357505
#> [6,]  0.1410282 -0.46923193
#> 
#> , , 70
#> 
#>              [,1]        [,2]
#> [1,]  0.659699245  0.08174020
#> [2,] -0.414386738  0.52190915
#> [3,] -0.004399052  0.65506674
#> [4,]  0.619058798  0.22343611
#> [5,] -0.077002100  0.04745094
#> [6,] -0.062434492 -0.48953209
#> 
#> , , 71
#> 
#>             [,1]        [,2]
#> [1,]  0.69456195  0.12506422
#> [2,] -0.36920842  0.53701484
#> [3,] -0.03374949  0.65932724
#> [4,]  0.61422854  0.19187409
#> [5,] -0.03527873  0.08242665
#> [6,] -0.04010634 -0.46653164
#> 
#> , , 72
#> 
#>             [,1]        [,2]
#> [1,]  0.72778805  0.10877460
#> [2,] -0.41697861  0.49816687
#> [3,]  0.02879620  0.58146573
#> [4,]  0.53715360  0.18204726
#> [5,]  0.08284868  0.05960483
#> [6,] -0.01504262 -0.60431905
#> 
#> , , 73
#> 
#>             [,1]        [,2]
#> [1,]  0.67481204  0.22954153
#> [2,] -0.18608310  0.27124224
#> [3,] -0.02682144  0.80933657
#> [4,]  0.69811450 -0.02037321
#> [5,] -0.04910691  0.09186425
#> [6,]  0.13966769 -0.45810320
#> 
#> , , 74
#> 
#>             [,1]        [,2]
#> [1,]  0.68591448  0.15057603
#> [2,] -0.35030731  0.54357194
#> [3,] -0.02869406  0.62999593
#> [4,]  0.63403890  0.16395096
#> [5,] -0.06216799  0.06569888
#> [6,] -0.01061074 -0.50375120
#> 
#> , , 75
#> 
#>              [,1]        [,2]
#> [1,]  0.708491753  0.19888528
#> [2,] -0.477593754  0.49074064
#> [3,]  0.096769352  0.72161201
#> [4,]  0.501480499  0.04745458
#> [5,] -0.095033952 -0.03585378
#> [6,]  0.008074172 -0.44199201
#> 
#> , , 76
#> 
#>             [,1]        [,2]
#> [1,]  0.86347398  0.15813644
#> [2,] -0.03813609  0.33496143
#> [3,] -0.16387531  0.49446325
#> [4,]  0.38585779  0.06113169
#> [5,]  0.24360931  0.14997181
#> [6,]  0.13368428 -0.76946161
#> 
#> , , 77
#> 
#>             [,1]       [,2]
#> [1,]  0.74295690  0.2869350
#> [2,] -0.46815338  0.6572972
#> [3,]  0.02125276  0.5696425
#> [4,]  0.46558396  0.0872985
#> [5,]  0.02892861  0.1597809
#> [6,] -0.10387726 -0.3577501
#> 
#> , , 78
#> 
#>             [,1]       [,2]
#> [1,]  0.64439632  0.1149449
#> [2,] -0.38650666  0.4616246
#> [3,] -0.08502445  0.6446317
#> [4,]  0.64299793  0.1463526
#> [5,]  0.06869792  0.1225546
#> [6,] -0.09985532 -0.5671874
#> 
#> , , 79
#> 
#>             [,1]        [,2]
#> [1,]  0.45531314  0.04033280
#> [2,] -0.50878377  0.60015627
#> [3,]  0.05441892  0.68203405
#> [4,]  0.63736366  0.40147833
#> [5,] -0.23364400  0.09181311
#> [6,] -0.26466141 -0.05831619
#> 
#> , , 80
#> 
#>             [,1]       [,2]
#> [1,]  0.59691832  0.1557599
#> [2,] -0.20118533  0.6951325
#> [3,]  0.11448651  0.3110501
#> [4,]  0.70534336  0.2612458
#> [5,]  0.07573214 -0.2049459
#> [6,]  0.29472226 -0.5343457
#> 
#> , , 81
#> 
#>             [,1]         [,2]
#> [1,]  0.66971552  0.248790355
#> [2,] -0.26134037  0.427898036
#> [3,] -0.02526891  0.660346962
#> [4,]  0.68419207  0.007442275
#> [5,] -0.09332274  0.005417753
#> [6,]  0.07560354 -0.564680245
#> 
#> , , 82
#> 
#>            [,1]       [,2]
#> [1,]  0.4902461  0.1405231
#> [2,] -0.3329224  0.7060531
#> [3,] -0.1435264  0.5725431
#> [4,]  0.7161717  0.1822523
#> [5,] -0.3112206 -0.3253928
#> [6,] -0.1358730 -0.1218210
#> 
#> , , 83
#> 
#>             [,1]       [,2]
#> [1,]  0.62087846  0.1845882
#> [2,] -0.44444717  0.5571341
#> [3,]  0.05931014  0.6461202
#> [4,]  0.63974026  0.1487471
#> [5,] -0.06111357  0.1602058
#> [6,] -0.02136572 -0.4361948
#> 
#> , , 84
#> 
#>            [,1]       [,2]
#> [1,]  0.6754549  0.2816900
#> [2,] -0.5199011  0.6371326
#> [3,]  0.1647252  0.5724495
#> [4,]  0.4773457  0.1024262
#> [5,]  0.1314560 -0.1222852
#> [6,] -0.0344877 -0.4019572
#> 
#> , , 85
#> 
#>             [,1]        [,2]
#> [1,]  0.73209218  0.17579206
#> [2,] -0.34484275  0.52795368
#> [3,] -0.04398543  0.61843192
#> [4,]  0.52989227  0.10696552
#> [5,] -0.24422768  0.01917815
#> [6,] -0.05250547 -0.54414575
#> 
#> , , 86
#> 
#>              [,1]        [,2]
#> [1,]  0.685729823  0.14813441
#> [2,] -0.363621557  0.53752867
#> [3,] -0.014067179  0.64434251
#> [4,]  0.627395167  0.16401002
#> [5,] -0.060527081  0.06635197
#> [6,] -0.008237988 -0.49258500
#> 
#> , , 87
#> 
#>            [,1]        [,2]
#> [1,]  0.6809473  0.22904771
#> [2,] -0.3612140  0.53936032
#> [3,]  0.1423364  0.64929729
#> [4,]  0.5631919  0.08393574
#> [5,] -0.2069873  0.13304851
#> [6,]  0.1598336 -0.45857770
#> 
#> , , 88
#> 
#>             [,1]        [,2]
#> [1,]  0.66555130  0.41024610
#> [2,] -0.30931710  0.47574771
#> [3,] -0.10978095  0.76685159
#> [4,]  0.65067589 -0.07738308
#> [5,] -0.04715736  0.07202685
#> [6,] -0.15397928 -0.07826163
#> 
#> , , 89
#> 
#>             [,1]        [,2]
#> [1,]  0.80375275  0.08714631
#> [2,] -0.13421890  0.48098817
#> [3,] -0.08093823  0.78722478
#> [4,]  0.50868317  0.21119062
#> [5,] -0.23465138  0.04595894
#> [6,]  0.12488382 -0.30760258
#> 
#> , , 90
#> 
#>             [,1]       [,2]
#> [1,]  0.74188828  0.1395184
#> [2,] -0.06095220  0.2771825
#> [3,] -0.08406232  0.8251135
#> [4,]  0.61064499  0.0272362
#> [5,] -0.14988325 -0.2903694
#> [6,]  0.20848942 -0.3712628
#> 
#> , , 91
#> 
#>             [,1]       [,2]
#> [1,]  0.68107606  0.1153822
#> [2,] -0.42589413  0.6013403
#> [3,]  0.14144735  0.7132823
#> [4,]  0.52804340  0.2293108
#> [5,] -0.21639931  0.1041556
#> [6,]  0.09530867 -0.2299421
#> 
#> , , 92
#> 
#>             [,1]       [,2]
#> [1,]  0.68728407  0.1121635
#> [2,] -0.37446559  0.5773396
#> [3,]  0.05640208  0.5960559
#> [4,]  0.61150316  0.2411275
#> [5,] -0.08325050  0.1882427
#> [6,]  0.05803594 -0.4530319
#> 
#> , , 93
#> 
#>             [,1]       [,2]
#> [1,]  0.67990890  0.1489284
#> [2,] -0.34968334  0.6038154
#> [3,] -0.06355208  0.6078383
#> [4,]  0.63834844  0.1965819
#> [5,] -0.03344907  0.0266440
#> [6,] -0.05290556 -0.4521124
#> 
#> , , 94
#> 
#>             [,1]        [,2]
#> [1,]  0.40356288  0.36305153
#> [2,] -0.39480736  0.55047652
#> [3,]  0.01486615  0.58540518
#> [4,]  0.72670755  0.09041662
#> [5,] -0.29525182  0.31043384
#> [6,] -0.25644815 -0.34340300
#> 
#> , , 95
#> 
#>             [,1]        [,2]
#> [1,]  0.66289094  0.23325373
#> [2,] -0.36492249  0.50319490
#> [3,]  0.04598939  0.65689084
#> [4,]  0.63575160  0.05306167
#> [5,] -0.13470753  0.05520473
#> [6,]  0.05446047 -0.50499397
#> 
#> , , 96
#> 
#>             [,1]       [,2]
#> [1,]  0.73002677  0.1983376
#> [2,] -0.37135101  0.4986825
#> [3,]  0.04871745  0.6016178
#> [4,]  0.52105462  0.1263617
#> [5,] -0.23483019  0.2594570
#> [6,] -0.01195050 -0.5164773
#> 
#> , , 97
#> 
#>             [,1]        [,2]
#> [1,]  0.68767701  0.15144358
#> [2,] -0.34328553  0.54930551
#> [3,] -0.04036822  0.64035692
#> [4,]  0.63586746  0.16700675
#> [5,] -0.05516374  0.06586073
#> [6,] -0.01597823 -0.48274469
#> 
#> , , 98
#> 
#>             [,1]        [,2]
#> [1,]  0.68914214  0.12941344
#> [2,] -0.16506304  0.34169139
#> [3,] -0.12091704  0.50383192
#> [4,]  0.69072828  0.09461452
#> [5,] -0.06246547  0.01122664
#> [6,]  0.04699883 -0.77690064
#> 
#> , , 99
#> 
#>             [,1]       [,2]
#> [1,]  0.76982473  0.1499471
#> [2,] -0.40072504  0.5047296
#> [3,] -0.04445787  0.6156059
#> [4,]  0.46117641  0.1330353
#> [5,] -0.13379081  0.1054356
#> [6,] -0.11928608 -0.5612293
#> 
#> , , 100
#> 
#>              [,1]       [,2]
#> [1,]  0.627023598  0.1423544
#> [2,] -0.400280869  0.6769950
#> [3,] -0.005995635  0.4051668
#> [4,]  0.485509111  0.4132403
#> [5,] -0.318063554  0.3352464
#> [6,] -0.331205622 -0.2722041
#> 
#> attr(,"class")
#> [1] "history_array"