Initialize anaconda environment used for python backend

Initialize anaconda environment used for python backend

Usage

init_env(env_name = "r-lionfish", virtual_env = "virtual_env", local = FALSE)

Arguments

env_name

a string that defines the name of the python environment reticulate uses. This can be useful if one wants to use a preinstalled python environment.

virtual_env

either "virtual_env" or "anaconda". "virtual_env" creates a virtual environment, which has the advantage that the GUI looks much nicer and no previous python installation is required,but the setup of the environment can be more error prone. "anaconda" installs the python environment via Anaconda, which can be more stable, but the GUI looks more dated.

Value

Examples

init_env(env_name="r-lionfish", virtual_env = "virtual_env")
#> Installing pyenv ...
#> Done! pyenv has been installed to '/home/runner/.local/share/r-reticulate/pyenv/bin/pyenv'.
#> + /home/runner/.local/share/r-reticulate/pyenv/bin/pyenv update
#> + /home/runner/.local/share/r-reticulate/pyenv/bin/pyenv install --skip-existing 3.10.12
#> Using Python: /home/runner/.pyenv/versions/3.10.12/bin/python3.10
#> Creating virtual environment 'r-lionfish' ... 
#> + /home/runner/.pyenv/versions/3.10.12/bin/python3.10 -m venv /home/runner/.virtualenvs/r-lionfish
#> Done!
#> Installing packages: pip, wheel, setuptools
#> + /home/runner/.virtualenvs/r-lionfish/bin/python -m pip install --upgrade pip wheel setuptools
#> Installing packages: numpy
#> + /home/runner/.virtualenvs/r-lionfish/bin/python -m pip install --upgrade --no-user numpy
#> Virtual environment 'r-lionfish' successfully created.
#> Using virtual environment 'r-lionfish' ...
#> + /home/runner/.virtualenvs/r-lionfish/bin/python -m pip install --upgrade --no-user pandas
#> Using virtual environment 'r-lionfish' ...
#> + /home/runner/.virtualenvs/r-lionfish/bin/python -m pip install --upgrade --no-user matplotlib
#> Using virtual environment 'r-lionfish' ...
#> + /home/runner/.virtualenvs/r-lionfish/bin/python -m pip install --upgrade --no-user customtkinter
#> Using virtual environment 'r-lionfish' ...
#> + /home/runner/.virtualenvs/r-lionfish/bin/python -m pip install --upgrade --no-user statsmodels
#> Using virtual environment 'r-lionfish' ...
#> + /home/runner/.virtualenvs/r-lionfish/bin/python -m pip install --upgrade --no-user seaborn
#> Python environment "r-lionfish" successfully loaded