R wrapper for the interactive_tour function written on python

R wrapper for the interactive_tour function written on python

Usage

interactive_tour(
  data,
  plot_objects,
  feature_names,
  half_range,
  n_plot_cols,
  preselection = FALSE,
  preselection_names = FALSE,
  n_subsets = 3,
  display_size = 5
)

Arguments

data

the dataset you want to investigate

plot_objects

a named list of objects you want to be displayed. Each entry requires a definition of the type of display and a specification of what should be plotted.

feature_names

names of the features of the dataset

half_range

factor that influences the scaling of the displayed tour plots. Small values lead to more spread out datapoints (that might not fit the plotting area), while large values lead to the data being more compact. If not provided a good estimate will be calculated and used.

n_plot_cols

specifies the number of columns of the grid of the final display.

preselection

a vector that specifies in which subset each datapoint should be put initially.

preselection_names

a vector that specifies the names of the preselection subsets

n_subsets

the total number of available subsets (up to 10).

display_size

rough size of each subplot in inches

Value

Examples